MUKEYS_by_ll_bbox {soilDB}	R Documentation

Fetch Map Unit Keys by Bounding Box

Description

This function will return a vector of map unit keys, from the Soil Data Access website, that correspond with a bounding box defined by WGS84 long/lat pairs.

Usage


MUKEYS_by_ll_bbox(bbox)

Arguments

bbox
a bounding box in WGS coordinates

Details

The SDA website can be found at http://sdmdataaccess.nrcs.usda.gov. See examples for bounding box formatting.

Value

a vector of map unit keys

Note

This function requires the 'RCurl' (http://www.omegahat.org/RCurl/) and 'XML' (http://www.omegahat.org/RSXML/) packages.

Author(s)

Dylan E Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

See Also

SDA_query, mapunit_geom_by_ll_bbox

Examples


# fetch map unit keys from the bounding-box:
# 
#         +------------- (-120.8, 37.8)
#         |                     |
#         |                     |
# (-120.95, 37.7) --------------+

b <- c(-120.9,37.7,-120.8,37.8)
## Not run: m <- MUKEYS_by_ll_bbox(b)


[Package soilDB version 0.9-5 Index]