ROSETTA Model Versions 1 and 3 API

A simple interface to the ROSETTA model for predicting hydraulic parameters from soil properties. The ROSETTA API was developed by Dr. Todd Skaggs (USDA-ARS) and links to the work of Zhang and Schaap, (2017).

ROSETTA(x, vars, v = c("1", "3"), conf = NULL)

Arguments

x
a data.frame of required soil properties, may contain other columns, see details

vars
character vector of soil properties in x containing relevant soil property values, see details

v
single character of '1' or '3', this is the model version number

conf
configuration passed to httr::POST().

Value

a data.frame object with estimated water retention curve parameters and saturated hydraulic conductivity:

...

all columns present in x

theta_r:

residual volumetric water content (cm^3/cm^3)

theta_s:

saturated volumetric water content (cm^3/cm^3)

alpha:

related to the inverse of the air entry suction, log10-tranformed values with units of cm

npar:

index of pore size distribution, log10-tranformed values with units of 1/cm

ksat:

saturated hydraulic conductivity, log10-transformed values with units of cm/day

Details

Soil properties supplied in x must be described, in order, via vars argument. The column names are irrelevant but must follow the ordering: sand, silt, clay, bulk density, volumetric water content at 33kPa, and volumetric water content at 1500kPa.

The ROSETTA model relies on a minimum of 3 soil properties, with increasing (expected) accuracy as additional properties are included:

required, sand, silt, clay: USDA soil texture separates (percentages) that sum to 100%

optional, bulk density (method?): mass per volume afte accounting for >2mm fragments, units of gm/cm3

optional, volumetric water content at 33 kPa: roughly "field capacity" for most soils, units of cm^3/cm^3

optional, volumetric water content at 1500 kPa: roughly "permanent wilting point" for most plants, units of cm^3/cm^3

References

Consider using the interactive version, with copy/paste functionality at: https://www.handbook60.org/rosetta.

Rosetta Model Home Page: https://www.ars.usda.gov/pacific-west-area/riverside-ca/agricultural-water-efficiency-and-salinity-research-unit/docs/model/rosetta-model/.

Python ROSETTA model: http://www.u.arizona.edu/~ygzhang/download.html.

Yonggen Zhang, Marcel G. Schaap. 2017. Weighted recalibration of the Rosetta pedotransfer model with improved estimates of hydraulic parameter distributions and summary statistics (Rosetta3). Journal of Hydrology. 547: 39-53. https://doi.org/10.1016/j.jhydrol.2017.01.004.

Kosugi, K. 1999. General model for unsaturated hydraulic conductivity for soils with lognormal pore-size distribution. Soil Sci. Soc. Am. J. 63:270-277.

Mualem, Y. 1976. A new model predicting the hydraulic conductivity of unsaturated porous media. Water Resour. Res. 12:513-522.

Schaap, M.G. and W. Bouten. 1996. Modeling water retention curves of sandy soils using neural networks. Water Resour. Res. 32:3033-3040.

Schaap, M.G., Leij F.J. and van Genuchten M.Th. 1998. Neural network analysis for hierarchical prediction of soil water retention and saturated hydraulic conductivity. Soil Sci. Soc. Am. J. 62:847-855.

Schaap, M.G., and F.J. Leij, 1998. Database Related Accuracy and Uncertainty of Pedotransfer Functions, Soil Science 163:765-779.

Schaap, M.G., F.J. Leij and M. Th. van Genuchten. 1999. A bootstrap-neural network approach to predict soil hydraulic parameters. In: van Genuchten, M.Th., F.J. Leij, and L. Wu (eds), Proc. Int. Workshop, Characterization and Measurements of the Hydraulic Properties of Unsaturated Porous Media, pp 1237-1250, University of California, Riverside, CA.

Schaap, M.G., F.J. Leij, 1999, Improved prediction of unsaturated hydraulic conductivity with the Mualem-van Genuchten, Submitted to Soil Sci. Soc. Am. J.

van Genuchten, M.Th. 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Sci. Am. J. 44:892-898.