Soil Data Access Query

Submit a query to the Soil Data Acccess (SDA) website in SQL, get the results as a dataframe.

SDA_query(q)
makeChunks(ids, size=100)
format_SQL_in_statement(x)

Arguments

q
a valid T-SQL query surrounded by double quotes

ids
vector of IDs for chunking, contents aren't used just length

size
target chunk size

x
character vector to be packed into an SQL `IN` statement

Details

The SDA website can be found at http://sdmdataaccess.nrcs.usda.gov and query examples can be found at http://sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx. A library of query examples can be found at https://nasis.sc.egov.usda.gov/NasisReportsWebSite/limsreport.aspx?report_name=SDA-SQL_Library_Home.

SSURGO (detailed soil survey) and STATSGO (generalized soil survey) data are stored together within SDA. This means that queries that don't specify an area symbol may result in a mixture of SSURGO and STATSGO records. See the examples below and the SDA Tutorial for details.

Value

A dataframe containing the results. NULL is retutned when queries result in 0 matches rows.

Note

This function requires the `httr`, `jsonlite`, and `XML` packages

See also

Examples

# \donttest{

  ## get SSURGO export date for all soil survey areas in California
  # there is no need to filter STATSGO 
  # because we are filtering on SSURGO areasymbols
  q <- "SELECT areasymbol, saverest FROM sacatalog WHERE areasymbol LIKE 'CA%';"
  x <- SDA_query(q)

#> single result set, returning a data.frame

  head(x)

#>   areasymbol              saverest
#> 1      CA011  9/16/2019 7:09:53 PM
#> 2      CA013 9/17/2019 12:03:38 AM
#> 3      CA021  9/16/2019 7:40:41 PM
#> 4      CA031  9/16/2019 7:46:42 PM
#> 5      CA033  9/16/2019 7:55:27 PM
#> 6      CA041  9/16/2019 7:56:58 PM

  ## get SSURGO component data associated with the 
  ## Amador series / major component only
  # this query must explicitly filter out STATSGO data
  q <- "SELECT cokey, compname, comppct_r FROM legend\n
    INNER JOIN mapunit mu ON mu.lkey = legend.lkey\n
    INNER JOIN component co ON mu.mukey = co.mukey\n
    WHERE legend.areasymbol != 'US' AND compname = 'Amador';"

  res <- SDA_query(q)

#> single result set, returning a data.frame

  str(res)

#> 'data.frame':	54 obs. of  3 variables:
#>  $ cokey    : int  17596519 17596830 17596961 17597155 17597159 17616756 17616750 17605873 17619646 17619657 ...
#>  $ compname : chr  "Amador" "Amador" "Amador" "Amador" ...
#>  $ comppct_r: int  45 3 3 7 25 76 10 3 45 3 ...
#>  - attr(*, "SDA_id")= chr "Table"

  ## get component-level data for a specific soil survey area (Yolo county, CA)
  # there is no need to filter STATSGO because the query contains
  # an implicit selection of SSURGO data by areasymbol
  q <- "SELECT \n
    component.mukey, cokey, comppct_r, compname, taxclname, \n
    taxorder, taxsuborder, taxgrtgroup, taxsubgrp \n
    FROM legend \n
    INNER JOIN mapunit ON mapunit.lkey = legend.lkey \n
    LEFT OUTER JOIN component ON component.mukey = mapunit.mukey \n
    WHERE legend.areasymbol = 'CA113' ;"

  res <- SDA_query(q)

#> single result set, returning a data.frame

  str(res)

#> 'data.frame':	608 obs. of  9 variables:
#>  $ mukey      : int  757748 757748 757748 757748 757749 757749 757749 757749 757749 757750 ...
#>  $ cokey      : int  17678670 17678671 17678672 17678673 17678674 17678675 17678676 17678677 17678678 17678679 ...
#>  $ comppct_r  : int  80 2 8 10 80 1 5 10 4 80 ...
#>  $ compname   : chr  "Scribner" "Unnamed" "Corbiere" "Vina" ...
#>  $ taxclname  : chr  "Fine-loamy, mixed, superactive, thermic Cumulic Endoaquolls" NA "Fine, mixed, superactive, thermic Cumulic Vertic Endoaquolls" "Coarse-loamy, mixed, superactive, thermic Pachic Haploxerolls" ...
#>  $ taxorder   : chr  "Mollisols" NA "Mollisols" "Mollisols" ...
#>  $ taxsuborder: chr  "Aquolls" NA "Aquolls" "Xerolls" ...
#>  $ taxgrtgroup: chr  "Endoaquolls" NA "Endoaquolls" "Haploxerolls" ...
#>  $ taxsubgrp  : chr  "Cumulic Endoaquolls" NA "Cumulic Vertic Endoaquolls" "Pachic Haploxerolls" ...
#>  - attr(*, "SDA_id")= chr "Table"

  ## get tabular data based on result from spatial query
  # there is no need to filter STATSGO because
  # SDA_Get_Mukey_from_intersection_with_WktWgs84() implies SSURGO
  #
  # requires raster and rgeos packages because raster is suggested
  # and rgeos is additional
  if(require(raster) & require(rgeos)) {
    # text -> bbox -> WKT
    # xmin, xmax, ymin, ymax
    b <- c(-120.9, -120.8, 37.7, 37.8)
    p <- writeWKT(as(extent(b), 'SpatialPolygons'))
    q <- paste0("SELECT mukey, cokey, compname, comppct_r FROM component \n
      WHERE mukey IN (SELECT DISTINCT mukey FROM\n
      SDA_Get_Mukey_from_intersection_with_WktWgs84('"
      , p, "')) ORDER BY mukey, cokey, comppct_r DESC")

    x <- SDA_query(q)
    str(x)
  }

#> Loading required package: raster

#> Warning: package 'raster' was built under R version 3.5.3

#> 
#> Attaching package: 'raster'

#> The following objects are masked from 'package:aqp':
#> 
#>     metadata, metadata<-

#> Loading required package: rgeos

#> Warning: package 'rgeos' was built under R version 3.5.3

#> rgeos version: 0.5-2, (SVN revision 621)
#>  GEOS runtime version: 3.6.1-CAPI-1.10.1 
#>  Linking to sp version: 1.3-2 
#>  Polygon checking: TRUE 

#> single result set, returning a data.frame

#> 'data.frame':	337 obs. of  4 variables:
#>  $ mukey    : int  462527 462527 462527 462541 462541 462541 462541 462554 462554 462554 ...
#>  $ cokey    : int  17625717 17625718 17625719 17625770 17625771 17625772 17625773 17625815 17625816 17625817 ...
#>  $ compname : chr  "Alamo" "Madera" "San Joaquin" "Chualar" ...
#>  $ comppct_r: int  85 10 5 85 5 5 5 85 5 10 ...
#>  - attr(*, "SDA_id")= chr "Table"

# }