SDA_query_features {soilDB}	R Documentation

Soil Data Access Spatial Query

Description

Iterate over Spatial* object features and submit spatial queries to the SDA web-service.

Usage

SDA_query_features(x, id='pedon_id')

Arguments

x
a Spatial* object with more than 1 feature, any defined coordinate system

id
the column name in x that contains a unique ID for each feature

Details

The SDA website can be found at http://sdmdataaccess.nrcs.usda.gov. A detailed vignette can be found at https://r-forge.r-project.org/scm/viewvc.php/*checkout*/docs/soilDB/SDA-tutorial.html?root=aqp.

Value

A dataframe containing the results.

Note

This function requires the 'httr', 'jsonlite', 'XML', and 'rgeos' packages

Author(s)

D.E. Beaudette

[Package soilDB version 1.6.3 Index]