Get SSURGO Data via Spatial Query

Get SSURGO Data via Spatial Query to SoilWeb

SoilWeb_spatial_query(
  bbox = NULL,
  coords = NULL,
  what = "mapunit",
  source = "soilweb"
)

Arguments

bbox

a bounding box in WGS84 geographic coordinates, see examples

coords

a coordinate pair in WGS84 geographic coordinates, see examples

what

data to query, currently ignored

source

the data source, currently ignored

Value

The data returned from this function will depend on the query style. See examples below.

Details

Data are currently available from SoilWeb. These data are a snapshot of the "official" data. The snapshot date is encoded in the "soilweb_last_update" column in the function return value. Planned updates to this function will include a switch to determine the data source: "official" data via USDA-NRCS servers, or a "snapshot" via SoilWeb.

Note

This function should be considered experimental; arguments, results, and side-effects could change at any time. SDA now supports spatial queries, consider using SDA_query_features instead.

Author

D.E. Beaudette

Examples

# \donttest{
if(requireNamespace("curl") &
    curl::has_internet()) {
    
    # query by bbox
    SoilWeb_spatial_query(bbox=c(-122.05, 37, -122, 37.05))
    
    # query by coordinate pair
    SoilWeb_spatial_query(coords=c(-121, 38))
}
#>   ogc_fid areasymbol musym  mukey soilweb_last_update dist_meters
#> 1 1493094      ca077   220 462112          2021-09-09 92.17456302
# }