fetchNASISLabData {soilDB}	R Documentation

Fetch lab data used site/horizon data from a PedonPC database.

Description

Fetch KSSL laboratory pedon/horizon layer data from a local NASIS database, return as a SoilProfileCollection object.

Usage

fetchNASISLabData()

Details

This function currently works only on Windows, and requires a 'nasis_local' ODBC connection.

Value

a SoilProfileCollection class object

Note

This fuction attempts to do most of the boilerplate work when extracting KSSL laboratory site/horizon data from a local NASIS database. Lab pedons that have errors in their horizonation are excluded from the returned object, however, their IDs are printed on the console. See getHzErrorsNASIS for a simple approach to identifying pedons with problematic horizonation.

Author(s)

J.M. Skovlin and D.E. Beaudette

See Also

get_labpedon_data_from_NASIS_db

Examples

## Not run: 
##D # query depends on some lab data, queried against the national database
##D # note that you must setup this connection ahead of time
##D # see inst/doc/setup_ODBC_local_NASIS.pdf
##D f <- fetchNASISLabData()
##D 
##D # plot only those profiles with densic contact
##D #plot(f[which(f$densic.contact), ], name='hzname')
##D 
## End(Not run)

[Package soilDB version 1.6.3 Index]