gSSURGO.chunk {soilDB}	R Documentation

Gridded SSURGO Chunk

Description

A chunk of the gridded SSURGO database (gSSURGO)

Usage

data(gSSURGO.chunk)

Details

This is a 106x137 grid, cropped from the gSSURGO database. Cell values are map unit keys (mukey), stored as integers. No raster attribute table (RAT) is included with this sample data set. Note that this sample of the gSSURGO data has been modified such that cell values are map unit keys, rather than the gSSURGO integer key.

Source

http://soils.usda.gov/survey/geography/ssurgo/description_gssurgo.html

Examples

data(gSSURGO.chunk)

[Package soilDB version 1.3-5 Index]