Low-level methods for getting projects and tables from NASIS Web Reports

get_NASISWebReport(): Gets a table from the "ReportData" section of specified report_name

get_NASISWebProjects(): Gets a table of project name, state, approval statu and project record ID for specified office, fiscal year and project name pattern.

get_NASISWebReport(report_name, proj_id, verbose = TRUE)

get_NASISWebProjects(msso, fy, project, verbose = TRUE)

Arguments

report_name
internal report name

proj_id
integer; Project record ID

verbose
Default: TRUE; print messages with report name and parameters?

msso
MLRA soil survey office e.g. "2-SON"

fy
Fiscal year e.g. 2020

project
Expression to match in project name e.g. "MLRA 18%"

Value

A data.frame, or (invisible) try-error if request fails