Extract component data from a local NASIS Database

Extract component data from a local NASIS Database.

get_component_data_from_NASIS_db(SS = TRUE, stringsAsFactors = default.stringsAsFactors())
get_component_restrictions_from_NASIS_db(SS = TRUE)

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Dylan E. Beaudette, Stephen Roecker, and Jay M. Skovlin

See also

Examples

# \donttest{
if(local_NASIS_defined()) {
 # query text note data
 fc <- try(get_component_data_from_NASIS_db())

 # show structure of component data returned
 str(fc)
}

#> 'data.frame':	1131 obs. of  57 variables:
#>  $ dmudesc         : chr  "025Cp" "025Cp" "025Cp" "025Fm" ...
#>  $ compname        : chr  "Colo" "Nodaway" "Zook" "Fillmore" ...
#>  $ comppct_r       : int  50 45 5 99 1 100 100 100 95 4 ...
#>  $ compkind        : chr  "series" "series" "series" "series" ...
#>  $ majcompflag     : int  1 1 0 1 0 1 1 1 1 0 ...
#>  $ localphase      : chr  "frequently flooded" "frequently flooded" NA "frequently ponded" ...
#>  $ drainagecl      : chr  "poorly" "moderately well" "poorly" "somewhat poorly" ...
#>  $ hydricrating    : chr  "yes" "no" "yes" "yes" ...
#>  $ elev_l          : num  244 244 244 244 244 244 244 244 244 244 ...
#>  $ elev_r          : num  305 305 305 305 305 305 305 305 305 305 ...
#>  $ elev_h          : num  396 396 396 396 396 396 396 396 396 396 ...
#>  $ slope_l         : num  0 0 0 0 0 6 0 6 0 0 ...
#>  $ slope_r         : num  1 1 1 0 0 13 1 9 1 1 ...
#>  $ slope_h         : num  2 2 2 1 0 20 2 11 2 2 ...
#>  $ aspectccwise    : int  0 0 0 0 0 0 0 0 0 0 ...
#>  $ aspectrep       : int  135 135 135 135 135 135 135 135 135 135 ...
#>  $ aspectcwise     : int  359 359 359 359 359 359 359 359 359 359 ...
#>  $ map_l           : int  762 762 762 762 NA 762 762 762 762 762 ...
#>  $ map_r           : int  787 787 787 787 NA 787 787 787 787 787 ...
#>  $ map_h           : int  813 813 813 813 NA 813 813 813 813 813 ...
#>  $ maat_l          : num  11 11 11 11 NA 11 11 11 11 11 ...
#>  $ maat_r          : num  12 12 12 12 NA 12 12 12 12 12 ...
#>  $ maat_h          : num  13 13 13 13 NA 13 13 13 13 13 ...
#>  $ mast_r          : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ reannualprecip_r: int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ffd_l           : int  160 160 160 160 NA 160 160 160 160 160 ...
#>  $ ffd_r           : int  170 170 170 170 NA 170 170 170 170 170 ...
#>  $ ffd_h           : int  180 180 180 180 NA 180 180 180 180 180 ...
#>  $ tfact           : int  5 5 5 3 5 1 5 5 5 5 ...
#>  $ wei             : chr  "48" "48" "86" "48" ...
#>  $ weg             : chr  "6" "6" "4" "6" ...
#>  $ nirrcapcl       : chr  "3" "3" "2" "4" ...
#>  $ nirrcapscl      : chr  "w" "w" "w" "w" ...
#>  $ nirrcapunit     : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ irrcapcl        : chr  NA NA NA "4" ...
#>  $ irrcapscl       : chr  NA NA NA "w" ...
#>  $ irrcapunit      : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ frostact        : chr  "high" "high" "high" "high" ...
#>  $ hydricrating.1  : int  1 2 1 1 1 2 2 2 2 1 ...
#>  $ hydgrp          : chr  "c/d" "b" "c/d" "d" ...
#>  $ corcon          : chr  "moderate" "low" "moderate" "moderate" ...
#>  $ corsteel        : chr  "high" "high" "high" "high" ...
#>  $ taxclname       : chr  "Fine-silty, mixed, mesic Cumulic Endoaquolls" "Fine-silty, mixed, superactive, nonacid, mesic Mollic Udifluvents" "Fine, smectitic, mesic Cumulic Vertic Endoaquolls" "Fine, smectitic, mesic Typic Argialbolls" ...
#>  $ taxorder        : chr  "mollisols" "entisols" "mollisols" "mollisols" ...
#>  $ taxsuborder     : chr  "aquolls" "fluvents" "aquolls" "albolls" ...
#>  $ taxgrtgroup     : chr  "endoaquolls" "udifluvents" "endoaquolls" "argialbolls" ...
#>  $ taxsubgrp       : chr  "cumulic endoaquolls" "mollic udifluvents" "cumulic vertic endoaquolls" "typic argialbolls" ...
#>  $ taxpartsize     : chr  "fine-silty" "fine-silty" "fine" "fine" ...
#>  $ taxpartsizemod  : chr  NA NA NA NA ...
#>  $ taxceactcl      : chr  NA "superactive" NA NA ...
#>  $ taxreaction     : chr  "not used" "nonacid" "not used" "not used" ...
#>  $ taxtempcl       : chr  "mesic" "mesic" "mesic" "mesic" ...
#>  $ taxmoistscl     : chr  NA NA NA NA ...
#>  $ taxtempregime   : chr  "mesic" "mesic" "mesic" "mesic" ...
#>  $ soiltaxedition  : chr  NA NA NA NA ...
#>  $ coiid           : int  146851 146852 146853 146854 1464912 146858 146860 146870 146883 146885 ...
#>  $ dmuiid          : int  69416 69416 69416 69417 69417 69420 69422 69430 69439 69439 ...

# }