Extract accessory tables and summaries from a local NASIS Database

Extract accessory tables and summaries from a local NASIS Database.

get_extended_data_from_NASIS_db(SS = TRUE, nullFragsAreZero = TRUE,
                                stringsAsFactors = default.stringsAsFactors()
                                )

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

nullFragsAreZero
should fragment volumes of NULL be interpreted as 0? (default: TRUE), see details

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Jay M. Skovlin and Dylan E. Beaudette

See also

Examples

# \donttest{

if(local_NASIS_defined()) {
 # query extended data
 e <- try(get_extended_data_from_NASIS_db())

 # show contents of extended data
 str(e)
}

#> Warning: some records are missing rock fragment volume, these have been removed

#> Warning: some records are missing artifact volume, these have been removed

#> Warning: all records are missing artifact volume (NULL). buffering result with NA. will be converted to zero if nullFragsAreZero = TRUE.

#> List of 15
#>  $ ecositehistory   :'data.frame':	5 obs. of  5 variables:
#>   ..$ siteiid        : int [1:5] 509137 888590 1437568 1437573 1501681
#>   ..$ ecositeid      : chr [1:5] "R060AY040SD" "R060AY040SD" "R060AY011SD" "R060AY040SD" ...
#>   ..$ ecositenm      : chr [1:5] "Clayey 16-18\" P.Z." "Clayey 16-18\" P.Z." "Clayey 13-16\" P.Z." "Clayey 16-18\" P.Z." ...
#>   ..$ ecositecorrdate: POSIXct[1:5], format: NA NA ...
#>   ..$ es_classifier  : chr [1:5] NA NA NA NA ...
#>  $ diagnostic       :'data.frame':	107 obs. of  4 variables:
#>   ..$ peiid   : int [1:107] 122291 122291 122292 122292 122293 122293 124299 124299 124300 124300 ...
#>   ..$ featkind: chr [1:107] "ochric epipedon" "cambic horizon" "ochric epipedon" "cambic horizon" ...
#>   ..$ featdept: int [1:107] 0 11 0 7 0 6 0 16 0 15 ...
#>   ..$ featdepb: int [1:107] 11 74 7 75 6 69 16 56 15 65 ...
#>  $ diagHzBoolean    :'data.frame':	44 obs. of  10 variables:
#>   ..$ peiid               : int [1:44] 122291 122292 122293 124299 124300 124301 124302 124303 124304 124305 ...
#>   ..$ ochric.epipedon     : logi [1:44] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ cambic.horizon      : logi [1:44] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ paralithic.contact  : logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ slickensides        : logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ salt.accumulations  : logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ paralithic.materials: logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ mollic.epipedon     : logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ secondary.carbonates: logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ argillic.horizon    : logi [1:44] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ restriction      :'data.frame':	45 obs. of  8 variables:
#>   ..$ peiid   : int [1:45] 500243 573109 573110 573111 657614 657616 657617 657618 657619 657620 ...
#>   ..$ resdept : int [1:45] 61 54 54 52 97 86 82 98 91 95 ...
#>   ..$ resdepb : int [1:45] NA 203 203 203 NA NA NA NA NA NA ...
#>   ..$ resthk_l: int [1:45] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ resthk_r: int [1:45] NA 149 149 151 NA NA NA NA NA NA ...
#>   ..$ resthk_h: int [1:45] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ reskind : chr [1:45] "bedrock, paralithic" "bedrock, paralithic" "bedrock, paralithic" "bedrock, paralithic" ...
#>   ..$ reshard : chr [1:45] "weakly" "weakly" "weakly" "weakly" ...
#>  $ frag_summary     :'data.frame':	337 obs. of  18 variables:
#>   ..$ phiid               : int [1:337] 514588 514589 514590 514591 514592 514593 514598 514599 514600 514601 ...
#>   ..$ fine_gravel         : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ cobbles             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ stones              : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct_nopf: num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct     : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ frag_summary_v2  :'data.frame':	356 obs. of  18 variables:
#>   ..$ phiid               : int [1:356] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ fine_gravel         : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ cobbles             : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ stones              : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct_nopf: num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct     : num [1:356] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ art_summary      :'data.frame':	337 obs. of  14 variables:
#>   ..$ phiid              : int [1:337] 514588 514589 514590 514591 514592 514593 514598 514599 514600 514601 ...
#>   ..$ art_fgr            : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_gr             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_cb             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_st             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_by             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_ch             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_fl             : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_unspecified    : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_art_pct      : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_cohesive  : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_penetrable: num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_innocuous : num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_persistent: num [1:337] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ surf_frag_summary:'data.frame':	88 obs. of  10 variables:
#>   ..$ peiid              : int [1:88] 105622 105623 105735 105736 120253 120795 122122 122291 122292 122293 ...
#>   ..$ surface_fgravel    : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_gravel     : num [1:88] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_cobbles    : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_stones     : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_boulders   : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_channers   : num [1:88] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_flagstones : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_paragravel : logi [1:88] NA NA NA NA NA NA ...
#>   ..$ surface_paracobbles: logi [1:88] NA NA NA NA NA NA ...
#>  $ texmodifier      :'data.frame':	256 obs. of  5 variables:
#>   ..$ peiid : int [1:256] 105622 105622 105622 105622 105622 105622 105623 105623 105623 105623 ...
#>   ..$ phiid : int [1:256] 514588 514589 514590 514591 514592 514593 514598 514599 514600 514601 ...
#>   ..$ phtiid: int [1:256] 503843 486205 486206 486207 486208 503844 503849 503850 503851 503852 ...
#>   ..$ seqnum: int [1:256] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ texmod: chr [1:256] NA NA NA NA ...
#>  $ geomorph         :'data.frame':	167 obs. of  7 variables:
#>   ..$ peiid       : int [1:167] 105622 105623 105735 105736 120253 120253 120795 120795 122122 122122 ...
#>   ..$ geomfmod    : chr [1:167] NA NA NA NA ...
#>   ..$ geomfname   : chr [1:167] "knoll" "knoll" "plains" "upland" ...
#>   ..$ geomfeatid  : int [1:167] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ existsonfeat: int [1:167] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ geomfiidref : int [1:167] 68 68 326 335 14303 335 335 14303 14299 393 ...
#>   ..$ geomftname  : chr [1:167] "landform" "landform" "landscape" "landscape" ...
#>  $ taxhistory       :'data.frame':	109 obs. of  20 variables:
#>   ..$ peiid         : int [1:109] 105622 105622 105623 105623 105735 105735 105736 105736 120253 120253 ...
#>   ..$ classdate     : POSIXct[1:109], format: "1954-10-06 00:00:00" "2007-07-23 20:35:12" ...
#>   ..$ classifier    : chr [1:109] NA NA NA NA ...
#>   ..$ classtype     : chr [1:109] "sampled as" "correlated" "correlated" "sampled as" ...
#>   ..$ taxonname     : chr [1:109] "Pierre" "Pierre" "Pierre" "Pierre" ...
#>   ..$ localphase    : logi [1:109] NA NA NA NA NA NA ...
#>   ..$ taxonkind     : chr [1:109] NA "series" "series" NA ...
#>   ..$ seriesstatus  : chr [1:109] NA NA NA NA ...
#>   ..$ taxpartsize   : chr [1:109] NA "very-fine" "very-fine" NA ...
#>   ..$ taxorder      : chr [1:109] NA "vertisols" "vertisols" NA ...
#>   ..$ taxsuborder   : chr [1:109] NA "torrerts" "torrerts" NA ...
#>   ..$ taxgrtgroup   : chr [1:109] NA "torrerts" "torrerts" NA ...
#>   ..$ taxsubgrp     : chr [1:109] NA "typic torrerts" "typic torrerts" NA ...
#>   ..$ soiltaxedition: chr [1:109] NA NA NA NA ...
#>   ..$ osdtypelocflag: int [1:109] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ taxmoistcl    : chr [1:109] NA "aridic (torric)" "aridic (torric)" NA ...
#>   ..$ taxtempregime : chr [1:109] NA "mesic" "mesic" NA ...
#>   ..$ taxfamother   : chr [1:109] NA NA NA NA ...
#>   ..$ psctopdepth   : int [1:109] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pscbotdepth   : int [1:109] NA NA NA NA NA NA NA NA NA NA ...
#>  $ photo            :'data.frame':	0 obs. of  4 variables:
#>   ..$ siteiid  : int(0) 
#>   ..$ recdate  : chr(0) 
#>   ..$ textcat  : chr(0) 
#>   ..$ imagepath: chr(0) 
#>  $ pm               :'data.frame':	81 obs. of  10 variables:
#>   ..$ siteiid     : int [1:81] 105752 105753 119051 120378 120547 120548 120549 122555 122556 122557 ...
#>   ..$ seqnum      : int [1:81] 1 1 2 2 2 2 2 2 2 2 ...
#>   ..$ pmorder     : int [1:81] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmdept      : int [1:81] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmdepb      : int [1:81] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmmodifier  : chr [1:81] NA NA NA NA ...
#>   ..$ pmgenmod    : logi [1:81] NA NA NA NA NA NA ...
#>   ..$ pmkind      : chr [1:81] "residuum" "residuum" "unconsolidated sediments" "residuum" ...
#>   ..$ pmorigin    : chr [1:81] "shale, unspecified" "shale, unspecified" "shale, unspecified" "shale, unspecified" ...
#>   ..$ pmweathering: chr [1:81] NA NA NA "slight" ...
#>  $ struct           :'data.frame':	270 obs. of  6 variables:
#>   ..$ phiid        : int [1:270] 514588 514588 514589 514589 514590 514590 514590 514591 514591 514591 ...
#>   ..$ structgrade  : chr [1:270] "weak" "strong" "moderate" "moderate" ...
#>   ..$ structsize   : chr [1:270] "medium" "very fine" "medium" "very fine" ...
#>   ..$ structtype   : chr [1:270] "subangular blocky" "granular" "prismatic" "angular blocky" ...
#>   ..$ structid     : int [1:270] 1 2 1 2 1 2 3 1 2 3 ...
#>   ..$ structpartsto: int [1:270] 2 NA 2 NA 2 3 NA 2 3 NA ...
#>  $ hzdesgn          :'data.frame':	337 obs. of  16 variables:
#>   ..$ phiid           : int [1:337] 514588 514589 514590 514591 514592 514593 514598 514599 514600 514601 ...
#>   ..$ seqnum          : int [1:337] 1 2 3 4 5 6 1 2 3 4 ...
#>   ..$ hzname          : chr [1:337] "A11" "A12" "B21" "B22" ...
#>   ..$ hzdept          : int [1:337] 0 5 13 25 46 71 0 4 10 30 ...
#>   ..$ hzdepb          : int [1:337] 5 13 25 46 71 97 4 10 30 51 ...
#>   ..$ desgndisc       : int [1:337] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ desgnmaster     : chr [1:337] NA NA NA NA ...
#>   ..$ desgnmasterprime: chr [1:337] NA NA NA NA ...
#>   ..$ desgnvert       : int [1:337] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ w               : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ r               : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ ss              : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ k               : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ z               : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ y               : logi [1:337] NA NA NA NA NA NA ...
#>   ..$ t               : logi [1:337] NA NA NA NA NA NA ...

# }