get_extended_data_from_NASIS_db {soilDB}	R Documentation

Extract accessory tables and summaries from a local NASIS Database

Description

Extract accessory tables and summaries from a local NASIS Database.

Usage


get_extended_data_from_NASIS_db()

Details

This function currently works only on Windows.

Value

A list with the results.

Author(s)

Jay M. Skovlin and Dylan E. Beaudette

See Also

get_hz_data_from_NASIS_db, get_site_data_from_NASIS_db

[Package soilDB version 1.2-8 Index]