Get accessory tables and summaries from a local pedonPC Database

Get accessory tables and summaries from a local pedonPC Database.

Usage

get_extended_data_from_pedon_db(dsn)

Arguments

dsn

The path to a 'pedon.mdb' database.

Value

A list with the results.

See also

Author

Jay M. Skovlin and Dylan E. Beaudette