Extract text note data from a local NASIS Database

Extract text note data from a local NASIS Database.

get_text_notes_from_NASIS_db(SS = TRUE, fixLineEndings = TRUE)

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

fixLineEndings
convert line endings from "\r\n" to "\n"

Details

This function currently works only on Windows.

Value

A list with the results.

See also

Examples

# \donttest{
# query text note data
t <- try(get_text_notes_from_NASIS_db())

# show contents text note data, includes: siteobs, site, pedon, horizon level text notes data.
str(t)

#> List of 5
#>  $ pedon_text  :'data.frame':	409 obs. of  8 variables:
#>   ..$ recdate      : POSIXct[1:409], format: "2008-04-04 00:00:00" "2012-04-06 00:00:00" ...
#>   ..$ recauthor    : chr [1:409] "LGH" "Tonie Endres" "Tonie Endres" "LGH" ...
#>   ..$ pedontextkind: Factor w/ 9 levels "pedon note, formatted",..: 2 3 3 2 3 2 6 6 2 2 ...
#>   ..$ textcat      : chr [1:409] "editnote" NA NA "editnote" ...
#>   ..$ textsubcat   : chr [1:409] NA NA NA NA ...
#>   ..$ textentry    : chr [1:409] "Site and pedon number may be different than number given to original description." "Ownership changed from Illinois 108A_108B Shared to 11-04 Aurora MLRA PO. This will need to be checked to deter"| __truncated__ "Ownership changed from Illinois 108A_108B Shared to 11-04 Aurora MLRA PO. This will need to be checked to deter"| __truncated__ "Site and pedon number may be different than the number given to original description." ...
#>   ..$ peiid        : int [1:409] 35313 35313 35317 35317 35318 35318 35404 35404 36266 36266 ...
#>   ..$ petextiid    : int [1:409] 206095 511532 511536 206116 511537 206099 67762 67893 203656 68661 ...
#>  $ site_text   :'data.frame':	511 obs. of  8 variables:
#>   ..$ recdate     : POSIXct[1:511], format: "2012-04-06 00:00:00" "2012-04-06 00:00:00" ...
#>   ..$ recauthor   : chr [1:511] "Tonie Endres" "Tonie Endres" "Tonie Endres" NA ...
#>   ..$ sitetextkind: Factor w/ 7 levels "site note, formatted",..: 3 3 3 6 5 6 5 2 3 5 ...
#>   ..$ textcat     : chr [1:511] NA NA NA "Map Unit Symbol/Name" ...
#>   ..$ textsubcat  : chr [1:511] NA NA NA NA ...
#>   ..$ textentry   : chr [1:511] "Ownership changed from Illinois 108A_108B Shared to 11-04 Aurora MLRA PO. This will need to be checked to deter"| __truncated__ "Ownership changed from Illinois 108A_108B Shared to 11-04 Aurora MLRA PO. This will need to be checked to deter"| __truncated__ "Ownership changed from Illinois 108A_108B Shared to 11-04 Aurora MLRA PO. This will need to be checked to deter"| __truncated__ "Map Unit Symbol: Wp\nMap Unit Name:   Westland silty clay loam" ...
#>   ..$ siteiid     : int [1:511] 35362 35366 35367 35458 35458 36322 36322 36322 36322 36387 ...
#>   ..$ sitetextiid : int [1:511] 495376 495380 495381 69688 69875 71081 71282 276574 495499 71163 ...
#>  $ siteobs_text:'data.frame':	540 obs. of  8 variables:
#>   ..$ recdate        : POSIXct[1:540], format: NA NA ...
#>   ..$ recauthor      : chr [1:540] NA NA NA NA ...
#>   ..$ siteobstextkind: Factor w/ 8 levels "site observation, formatted",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ textcat        : chr [1:540] NA NA NA NA ...
#>   ..$ textsubcat     : chr [1:540] NA NA NA NA ...
#>   ..$ textentry      : chr [1:540] NA NA NA NA ...
#>   ..$ site_id        : int [1:540] 35362 35366 35367 35458 36322 36387 36388 37117 38282 38283 ...
#>   ..$ siteobstextiid : int [1:540] NA NA NA NA NA NA NA NA NA NA ...
#>  $ horizon_text:'data.frame':	579 obs. of  8 variables:
#>   ..$ recdate         : POSIXct[1:579], format: "2008-03-25" "2008-02-06" ...
#>   ..$ recauthor       : chr [1:579] "LGH" "LGH" "LGH" NA ...
#>   ..$ phorizontextkind: Factor w/ 6 levels "horizon note, formatted",..: 2 3 2 2 2 2 2 2 2 2 ...
#>   ..$ textcat         : chr [1:579] "editnote" "editnote" "editnote" NA ...
#>   ..$ textsubcat      : chr [1:579] NA NA NA NA ...
#>   ..$ textentry       : chr [1:579] "Calculated textural modifier and class in pedon horizon." "Calculated textural modifier and class in pedon horizon." "Calculated textural modifier and class in pedon horizon." "median penetrometer reading of 0.5" ...
#>   ..$ phiid           : int [1:579] 166850 166896 167033 177763 177764 177793 177796 177797 177855 177856 ...
#>   ..$ phtextiid       : int [1:579] 248751 244731 245520 18716 18717 18718 18719 18720 18721 18722 ...
#>  $ photo_links :'data.frame':	0 obs. of  8 variables:
#>   ..$ recdate        : chr(0) 
#>   ..$ recauthor      : chr(0) 
#>   ..$ siteobstextkind: Factor w/ 8 levels "site observation, formatted",..: 
#>   ..$ textcat        : chr(0) 
#>   ..$ textsubcat     : chr(0) 
#>   ..$ textentry      : chr(0) 
#>   ..$ site_id        : int(0) 
#>   ..$ siteobstextiid : int(0) 

# view text categories for site text notes
if(!inherits(t, 'try-error'))
  table(t$site_text$textcat)

#> 
#>                          associated soils           assocsoils 
#>                   70                   10                   11 
#>              Caption           edit notes             editnote 
#>                    4                    1                    7 
#>                edits                GA259                GA639 
#>                    4                    2                    1 
#>                IL143             landform             Landform 
#>                    2                   14                   12 
#>             location             Location Map Unit Symbol/Name 
#>                   21                   21                   36 
#>         mini profile        not converted               noteid 
#>                    1                    3                    6 
#>     old pedon number    Pedon Description         Physiography 
#>                    1                    1                    1 
#>      quad correction                 RaCA              Site ID 
#>                    6                    2                    1 
#>                slope     soil macro fauna               Update 
#>                   19                    1                    1 
#>                   US                VA023                VA173 
#>                    1                    4                    1 
#>           vegetation 
#>                    3 

# }