Fetch Map Unit Geometry from SDA

Fetch map unit geometry from the SDA website by WGS84 bounding box. There is a limit on the amount of data returned as serialized JSON (~32Mb) and a total record limit of 100,000.

mapunit_geom_by_ll_bbox(bbox, source = "sda")

Arguments

bbox

a bounding box in WGS coordinates

source

the source database, currently limited to soil data access (SDA)

Value

A SpatialPolygonsDataFrame of map unit polygons, in WGS84 (long,lat) coordinates.

Details

The SDA website can be found at https://sdmdataaccess.nrcs.usda.gov. See examples for bounding box formatting.

Note

SDA does not return the spatial intersection of map unit polygons and bounding box. Rather, just those polygons that are completely within the bounding box / overlap with the bbox. This function requires the 'rgdal' package.

Author

Dylan E. Beaudette