seriesExtent {soilDB}	R Documentation

Get/Display Soil Series Extent

Description

Get or display the spatial extent of a named soil series using the SoilWeb query facility.

Usage


seriesExtent(s, timeout=60)
seriesExtentAsGmap(s, timeout=60, exp=1.25)

Arguments

s
the soil series name

timeout
time that we are willing to wait for a response, in seconds

exp
expansion factor used to expand Google Maps region

Details

These functions are only useful for querying the

Value

when calling seriesExtent, a SpatialPolygonsDataFrame object

Note

These function require the 'rgdal' and 'dismo' packages.

Author(s)

D. E. Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

Examples


## Not run: 
##D # fetch series extent for the 'Amador' soil series
##D s <- seriesExtent('amador')
##D plot(s)
##D  
##D # fetch then plot the extent of the 'Amador' soil series
##D seriesExtentAsGmap('amador')
## End(Not run)



[Package soilDB version 0.9-3 Index]