Lookup siblings and cousins for a given soil series.

Lookup siblings and cousins for a given soil series, from the current fiscal year SSURGO snapshot via SoilWeb.

siblings(s, only.major=FALSE, component.data = FALSE, cousins = FALSE)

Arguments

s
character vector, the name of a single soil series, case-insensitive.

only.major
logical, should only return siblings that are major components

component.data
logical, should component data for siblings (and optionally cousins) be returned?

cousins
logical, should siblings-of-siblings (cousins) be returned?

Details

The siblings of any given soil series are defined as those soil series (major and minor component) that share a parent map unit with the named series (as a major component). Cousins are siblings of siblings. Data are sourced from SoilWeb which maintains a copy of the current SSURGO snapshot.

Value

sib

data.frame containing siblings, major component flag, and number of co-occurrences

sib.data

data.frame containing sibling component data

cousins

data.frame containing cousins, major component flag, and number of co-occurrences

cousin.data

data.frame containing cousin component data

References

soilDB Soil Series Query Functionality

Related tutorial.

See also

Examples

# basic usage
x <- siblings('zook')
x$sib

#>    series      sibling majcompflag  n
#> 1    zook       Olmitz        TRUE 14
#> 2    zook        Vigar        TRUE 12
#> 3    zook       Vesser        TRUE 10
#> 4    zook          Ely        TRUE 10
#> 5    zook      Excello        TRUE  9
#> 6    zook         Colo        TRUE  8
#> 7    zook      Nodaway        TRUE  5
#> 8    zook Mt. Sterling        TRUE  3
#> 9    zook          Zoe        TRUE  3
#> 10   zook        Kezan        TRUE  2
#> 11   zook        Clamo        TRUE  2
#> 12   zook       Quiver        TRUE  1
#> 13   zook     Humeston        TRUE  1
#> 14   zook         Klum        TRUE  1
#> 15   zook       Wabash       FALSE 64
#> 16   zook         Colo       FALSE 53
#> 17   zook     Chequest       FALSE 46
#> 18   zook      Nodaway       FALSE 43
#> 19   zook     Humeston       FALSE 27
#> 20   zook       Arbela       FALSE 26
#> 21   zook      Ackmore       FALSE 18
#> 22   zook       Bremer       FALSE 16
#> 23   zook      Dockery       FALSE 16
#> 24   zook         Lamo       FALSE 15
#> 25   zook       Landes       FALSE 13
#> 26   zook         Napa       FALSE 12
#> 27   zook        Shell       FALSE 12
#> 28   zook       Olmitz       FALSE 11
#> 29   zook       Coland       FALSE 10
#> 30   zook      Sawmill       FALSE  9
#> 31   zook     Kennebec       FALSE  9
#> 32   zook       Vesser       FALSE  7
#> 33   zook       Nishna       FALSE  5
#> 34   zook     Blackoar       FALSE  2
#> 35   zook       Judson       FALSE  2
#> 36   zook      Reading       FALSE  2
#> 37   zook        Calco       FALSE  2
#> 38   zook        Chase       FALSE  2
#> 39   zook       Quiver       FALSE  2
#> 40   zook         Muir       FALSE  1
#> 41   zook       Floris       FALSE  1
#> 42   zook          Ely       FALSE  1
#> 43   zook     Clarinda       FALSE  1
#> 44   zook       Eudora       FALSE  1
#> 45   zook   Toolesboro       FALSE  1

# restrict to siblings that are major components
# e.g. the most likely siblings
x <- siblings('zook', only.major = TRUE)
x$sib

#>    series      sibling majcompflag  n
#> 1    zook       Olmitz        TRUE 14
#> 2    zook        Vigar        TRUE 12
#> 3    zook       Vesser        TRUE 10
#> 4    zook          Ely        TRUE 10
#> 5    zook      Excello        TRUE  9
#> 6    zook         Colo        TRUE  8
#> 7    zook      Nodaway        TRUE  5
#> 8    zook          Zoe        TRUE  3
#> 9    zook Mt. Sterling        TRUE  3
#> 10   zook        Kezan        TRUE  2
#> 11   zook        Clamo        TRUE  2
#> 12   zook     Humeston        TRUE  1
#> 13   zook       Quiver        TRUE  1
#> 14   zook         Klum        TRUE  1