SDA_query {soilDB}	R Documentation

Soil Data Access Query

Description

Submit a query to the Soil Data Acccess (SDA) website in SQL, get the results as a dataframe.

Usage

SDA_query(q)

Arguments

q
a text string containing valid SQL

Details

The SDA website can be found at http://sdmdataaccess.nrcs.usda.gov and query examples can be found at http://sdmdataaccess.nrcs.usda.gov/QueryHelp.aspx

Value

A dataframe containing the results.

Note

This function requires the 'SSOAP' (http://www.omegahat.org/SSOAP/), 'RCurl' (http://www.omegahat.org/RCurl/), 'XML' (http://www.omegahat.org/RSXML/), and 'XMLSchema' (http://www.omegahat.org/XMLSchema/) packages. Some Windows binary packages can be found here http://www.stats.ox.ac.uk/pub/RWin/bin/windows/contrib/2.15/. This function has been successfully tested on all Windows, MacOS, and Linux using SSOAP 0.8-1, XMLSchema 0.6-0, and XML 3.7-1.

Author(s)

Dylan E. Beaudette

References

http://casoilresource.lawr.ucdavis.edu/

See Also

MUKEYS_by_ll_bbox, mapunit_geom_by_ll_bbox

Examples

# install the required packages
## Not run: 
##D install.packages('RCurl', dep=TRUE)
##D install.packages('XML', dep=TRUE)
##D install.packages("SSOAP", repos = "http://www.omegahat.org/R", type="source")
##D install.packages("XMLSchema", repos = "http://www.omegahat.org/R", type="source")
## End(Not run)

# SSURGO export metadata:
## Not run: 
##D q <- "SELECT areasymbol, saverest FROM sacatalog WHERE areasymbol LIKE 'CA%';"
##D x <- SDA_query(q)
##D x$saverest <- as.Date(x$saverest, format="%m/%d/%Y")
##D head(x)
## End(Not run)

# basic query:
## Not run: 
##D res <- SDA_query("select cokey, compname, comppct_r 
##D from component 
##D where compname = 'yolo' and majcompflag = 'Yes' ")
## End(Not run)

# get tabular data based on result from spatial query:
## Not run: 
##D # bbox
##D b <- c(-120.9,37.7,-120.8,37.8)
##D # get map unit keys for this bbox
##D m <- MUKEYS_by_ll_bbox(b)
##D # make an SQL-compliant "in" statment
##D in.statement <- format_SQL_in_statement(m)
##D # format query:
##D q <- paste("
##D SELECT component.cokey, compname, comppct_r, hzdept_r, hzdepb_r, 
##D hzname, sandtotal_r, silttotal_r, claytotal_r
##D FROM component JOIN chorizon ON component.cokey = chorizon.cokey
##D WHERE majcompflag = 'Yes' AND mukey IN ", in.statement, 
##D "ORDER BY cokey, comppct_r DESC, hzdept_r ASC", sep="")
##D # now get component and horizon-level data for these map unit keys
##D res <- SDA_query(q)
## End(Not run)

# get component-level data for a specific soil survey area (Yolo county, CA)
## Not run: 
##D q <- "SELECT 
##D component.mukey, cokey, comppct_r, compname, taxclname, 
##D taxorder, taxsuborder, taxgrtgroup, taxsubgrp
##D FROM legend
##D INNER JOIN mapunit ON mapunit.lkey = legend.lkey
##D LEFT OUTER JOIN component ON component.mukey = mapunit.mukey
##D WHERE legend.areasymbol = 'CA113'"
##D 
##D res <- SDA_query(q)
## End(Not run)

[Package soilDB version 1.3-5 Index]