SoilWeb_spatial_query {soilDB}	R Documentation

Get SSURGO Data via Spatial Query

Description

Get SSURGO Data via Spatial Query to SoilWeb

Usage


SoilWeb_spatial_query(bbox = NULL, coords = NULL, what = "mapunit", source = "soilweb")

Arguments

bbox
a bounding box in WGS84 geographic coordinates, see examples

coords
a coordinate pair in WGS84 geographic coordinates, see examples

what
data to query, currently ignored

source
the data source, currently ignored

Details

Data are currently available from SoilWeb. These data are a snapshot of the "official" data. The snapshot date is encoded in the "soilweb_last_update" column in the function return value. Planned updates to this function will include a switch to determine the data source: "official" data via USDA-NRCS servers, or a "snapshot" via SoilWeb.

Value

The data returned from this function will depend on the query style. See examples below.

Note

This function should be considered experimental; arguments, results, and side-effects could change at any time. SDA now supports spatial queries, consider using SDA_query_features instead.

Author(s)

D.E. Beaudette

Examples

# query by bbox
## Not run: SoilWeb_spatial_query(bbox=c(-122.05, 37, -122, 37.05))

# query by coordinate pair
## Not run: SoilWeb_spatial_query(coords=c(-121, 38))

[Package soilDB version 1.6.3 Index]