fetchKSSL {soilDB}	R Documentation

Fetch KSSL Data (EXPERIMENTAL)

Description

Get KSSL data via BBOX, MLRA, or series name query, from the SoilWeb system.

Usage

fetchKSSL(series = NULL, bbox = NULL, mlra=NULL)

Arguments

series
a soil series name

bbox
a bounding box in WGS84 geographic coordinates e.g. c(-120, 37, -122, 38)

mlra
an MLRA ID

Details

This is an experimental interface to most of the KSSL data as of June 2015. Series-queries are case insensitive. Series name is based on the "correlated as" field (from KSSL snapshot) when present. The "sampled as" classification was promoted to "correlated as" if the "correlated as" classification was missing.

Value

a SoilProfileCollection object

Note

SoilWeb maintains a snapshot of the KSSL data. Please use the link below for the live data.

Author(s)

D.E. Beaudette

References

http://ncsslabdatamart.sc.egov.usda.gov/

See Also

fetchOSD

Examples

## Not run: 
##D # search by series name
##D s <- fetchKSSL(series='auburn')
##D 
##D # search by bounding-box
##D # s <- fetchKSSL(bbox=c(-120, 37, -122, 38))
##D 
##D # how many pedons
##D length(s)
##D 
##D # plot 
##D par(mar=c(0,0,0,0))
##D plot(s, name='hzn_desgn', max.depth=150)
## End(Not run)

[Package soilDB version 1.6.3 Index]