fetchNASIS {soilDB}	R Documentation

Fetch commonly used site/horizon data from a PedonPC database.

Description

Fetch commonly used site/horizon data from a PedonPC or local NASIS database, return as a SoilProfileCollection object.

Usage


fetchNASIS()
fetchNASIS_component_data()
getHzErrorsNASIS(strict=TRUE)

Arguments

strict
should horizonation by strictly enforced? (TRUE)

Details

This function currently works only on Windows, and requires a 'nasis_local' ODBC connection.

Value

a SoilProfileCollection class object

Note

This fuction attempts to do most of the boilerplate work when extracting site/horizon data from a local NASIS database. Pedons that have errors in their horizonation are excluded from the returned object, however, their IDs are printed on the console. See getHzErrorsNASIS for a simple approach to identifying pedons with problematic horizonation.

Author(s)

D. E. Beaudette and J. M. Skovlin

See Also

get_hz_data_from_pedon_db

Examples


## Not run: 
##D # query depends on some pedon data, queried against the national database
##D # note that you must setup this connection ahead of time
##D # see inst/doc/setup_ODBC_local_NASIS.pdf
##D f <- fetchNASIS()
##D 
##D # plot only those profiles with densic contact
##D plot(f[which(f$densic.contact), ], name='hzname')
##D 
##D # get basic component data from local NASIS, after performing a 
##D # DMU-* query against the national database
##D fc <- fetchNASIS_component_data()
## End(Not run)


[Package soilDB version 1.2-8 Index]