Get component tables from NASIS Web Reports

Get component tables from NASIS Web Reports

fetchNASISWebReport(
  projectname,
  rmHzErrors = FALSE,
  fill = FALSE,
  stringsAsFactors = default.stringsAsFactors()
)

get_component_from_NASISWebReport(
  projectname,
  stringsAsFactors = default.stringsAsFactors()
)

get_chorizon_from_NASISWebReport(
  projectname,
  fill = FALSE,
  stringsAsFactors = default.stringsAsFactors()
)

get_legend_from_NASISWebReport(
  mlraoffice,
  areasymbol,
  droplevels = TRUE,
  stringsAsFactors = default.stringsAsFactors()
)

get_lmuaoverlap_from_NASISWebReport(
  areasymbol,
  droplevels = TRUE,
  stringsAsFactors = default.stringsAsFactors()
)

get_mapunit_from_NASISWebReport(
  areasymbol,
  droplevels = TRUE,
  stringsAsFactors = default.stringsAsFactors()
)

get_projectmapunit_from_NASISWebReport(
  projectname,
  stringsAsFactors = default.stringsAsFactors()
)

get_projectmapunit2_from_NASISWebReport(
  mlrassoarea,
  fiscalyear,
  projectname,
  stringsAsFactors = default.stringsAsFactors()
)

get_project_from_NASISWebReport(mlrassoarea, fiscalyear)

get_progress_from_NASISWebReport(mlrassoarea, fiscalyear, projecttypename)

get_project_correlation_from_NASISWebReport(
  mlrassoarea,
  fiscalyear,
  projectname
)

Arguments

projectname

text string vector of project names to be inserted into a SQL WHERE clause (default: NA)

rmHzErrors

should pedons with horizonation errors be removed from the results? (default: FALSE)

fill

should rows with missing component ids be removed (default: FALSE)

stringsAsFactors

logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

mlraoffice

text string value identifying the MLRA Regional Soil Survey Office group name inserted into a SQL WHERE clause (default: NA)

areasymbol

text string value identifying the area symbol (e.g. IN001 or IN%) inserted into a SQL WHERE clause (default: NA) NULL (default: TRUE)

droplevels

logical: indicating whether to drop unused levels in classifying factors. This is useful when a class has large number of unused classes, which can waste space in tables and figures.

mlrassoarea

text string value identifying the MLRA Soil Survey Office areasymbol symbol inserted into a SQL WHERE clause (default: NA)

fiscalyear

text string value identifying the fiscal year inserted into a SQL WHERE clause (default: NA)

projecttypename

text string value identifying the project type name inserted into a SQL WHERE clause (default: NA)

Value

A data.frame or list with the results.

Author

Stephen Roecker