fetchOSD {soilDB}	R Documentation

Fetch Official Series Description (OSD) Data

Description

This functions fetches a limited subset of horizon and site-level attributes for named soil series, from the SoilWeb system.

Usage


fetchOSD(soils)

Arguments

soils
a character vector of named soil series

Details

the search is case-insensitive

Value

a SoilProfileCollection object

Note

SoilWeb maintains a snapshot of the Official Series Description data. Please use the link above for the live data.

Author(s)

D.E. Beaudette

References

http://soils.usda.gov/technical/classification/osd/index.html

Examples

## Not run: 
##D # soils of interest
##D s.list <- c('musick', 'cecil', 'drummer', 'amador', 'pentz', 
##D 'reiff', 'san joaquin', 'montpellier', 'grangeville', 'pollasky', 'ramona')
##D 
##D # fetch and convert data into an SPC
##D s <- fetchOSD(s.list)
##D 
##D # plot profiles
##D par(mar=c(0,0,0,0))
##D plot(s, name='hzname', cex.names=0.85, axis.line.offset=-4)
## End(Not run)

[Package soilDB version 1.3-5 Index]