Fetch KSSL Data (EXPERIMENTAL)

Get Rapid Carbon Assessment (RaCA) data via state, geographic bounding-box, RaCA site ID, or series query from the SoilWeb system.

fetchRaCA(series = NULL, bbox = NULL, state = NULL, rcasiteid = NULL, get.vnir = FALSE)

Arguments

series
a soil series name, case insensitive

bbox
a bounding box in WGS84 geographic coordinates e.g. c(-120, 37, -122, 38), constrained to a 5-degree block

state
a two-letter US state abbreviation, case insensitive

rcasiteid
an RaCA site id (e.g. 'C1609C01')

get.vnir
boolean, should associated VNIR spectra be downloaded? (see details)

Value

pedons:

a SoilProfileCollection object containing site/pedon/horizon data

trees:

a data.frame object containing tree DBH and height

veg:

a data.frame object containing plant species

stock:

a data.frame object containing carbon quantities (stocks) at standardized depths

sample:

a data.frame object containing sample-level bulk density and soil organic carbon values

spectra:

a numeric matrix containing VNIR reflectance spectra from 350--2500 nm

Details

The VNIR spectra associated with RaCA data are quite large [each gzip-compressed VNIR spectra record is about 6.6kb], so requests for these data are disabled by default. Note that VNIR spectra can only be queried by soil series or geographic BBOX.

References

http://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/?cid=nrcs142p2_054164 fetchRaCA() Tutorial

See also

Examples

# NOT RUN {
  # search by series name
  s <- fetchRaCA(series='auburn')

  # search by bounding-box
  # s <- fetchRaCA(bbox=c(-120, 37, -122, 38))

  # check structure
  str(s, 1)

  # extract pedons
  p <- s$pedons

  # how many pedons
  length(p)

  # plot 
  par(mar=c(0,0,0,0))
  plot(p, name='hzn_desgn', max.depth=150)
# }