Get Rapid Carbon Assessment (RaCA) data

Get Rapid Carbon Assessment (RaCA) data by state, geographic bounding-box, RaCA site ID, or soil series query from the SoilWeb API. This interface to the data was an experimental delivery service that does not include the latest soil carbon measurements.

Please use current RaCA distribution if you need measured SOC.

This interface will be updated sometime calendar year 2022 to include the latest soil morphology, taxonomic classification, and measured SOC values. More detailed coordinates for sample sites should also be available.

See https://www.nrcs.usda.gov/Internet/FSE_MANUSCRIPTS/alabama/raca_download/raca_download.zip for direct download of the full dataset.

fetchRaCA(
  series = NULL,
  bbox = NULL,
  state = NULL,
  rcasiteid = NULL,
  get.vnir = FALSE
)

Arguments

series

a soil series name; case-insensitive

bbox

a bounding box in WGS84 geographic coordinates e.g. c(-120, 37, -122, 38), constrained to a 5-degree block

state

a two-letter US state abbreviation; case-insensitive

rcasiteid

a RaCA site id (e.g. 'C1609C01')

get.vnir

logical, should associated VNIR spectra be downloaded? (see details)

Value

pedons:

a SoilProfileCollection object containing site/pedon/horizon data

trees:

a data.frame object containing tree DBH and height

veg:

a data.frame object containing plant species

stock:

a data.frame object containing carbon quantities (stocks) at standardized depths

sample:

a data.frame object containing sample-level bulk density and soil organic carbon values

spectra:

a numeric matrix containing VNIR reflectance spectra from 350--2500 nm

Details

The VNIR spectra associated with RaCA data are quite large (each gzip-compressed VNIR spectra record is about 6.6kb), so requests for these data are disabled by default. Note that VNIR spectra can only be queried by soil series or geographic BBOX.

References

https://www.nrcs.usda.gov/wps/portal/nrcs/detail/soils/survey/?cid=nrcs142p2_054164

See also

Author

D.E. Beaudette, USDA-NRCS staff