Get Daily Climate Data from USDA-NRCS SCAN (Soil Climate Analysis Network) Stations

Query soil/climate data from USDA-NRCS SCAN Stations

fetchSCAN(
  site.code = NULL,
  year = NULL,
  report = "SCAN",
  timeseries = c("Daily", "Hourly"),
  ...
)

SCAN_sensor_metadata(site.code)

SCAN_site_metadata(site.code = NULL)

Arguments

site.code

a vector of site codes. If NULL SCAN_site_metadata() returns metadata for all SCAN sites.

year

a vector of years

report

report name, single value only

timeseries

either 'Daily' or 'Hourly'

...

additional arguments. May include intervalType, format, sitenum, interval, year, month. Presence of additional arguments bypasses default batching functionality provided in the function and submits a 'raw' request to the API form.

Value

a data.frame object; NULL on bad request.

Details

Possible above and below ground sensor types include: 'SMS' (soil moisture), 'STO' (soil temperature), 'SAL' (salinity), 'TAVG' (daily average air temperature), 'TMIN' (daily minimum air temperature), 'TMAX' (daily maximum air temperature), 'PRCP' (daily precipitation), 'PREC' (daily precipitation), 'SNWD' (snow depth), 'WTEQ' (snow water equivalent),'WDIRV' (wind direction), 'WSPDV' (wind speed), 'LRADT' (solar radiation/langley total).

SCAN Sensors

All Soil Climate Analysis Network (SCAN) sensor measurements are reported hourly.

Element Measured	Sensor Type	Precision

Air Temperature	Shielded thermistor	0.1 degrees C

Barometric Pressure	Silicon capacitive pressure sensor	1%

Precipitation	Storage-type gage or tipping bucket	Storage: 0.1 inches;

Relative Humidity	Thin film capacitance-type sensor	1%

Snow Depth	Sonic sensor (not on all stations)	0.5 inches

Snow Water Content	Snow pillow device and a pressure transducer (not on all stations)	0.1 inches

Soil Moisture	Dielectric constant measuring device. Typical measurements are at 2", 4", 8", 20", and 40" where possible.	0.50%

Soil Temperature	Encapsulated thermistor. Typical measurements are at 2", 4", 8", 20", and 40" where possible.	0.1 degrees C

Solar Radiation	Pyranometer	0.01 watts per meter

Wind Speed and Direction	Propellor-type anemometer	Speed: 0.1 miles per hour; Direction: 1 degree

SNOTEL Sensors

All Snow Telemetry (SNOTEL) sensor measurements are reported daily.

Element Measured	Sensor Type	Precision

Air Temperature	Shielded thermistor	0.1 degrees C

Barometric Pressure	Silicon capacitive pressure sensor	1%

Precipitation	Storage-type gage or tipping bucket	Storage: 0.1 inches; Tipping bucket: 0.01 inches

Relative Humidity	Thin film capacitance-type sensor	1%

Snow Depth	Sonic sensor	0.5 inches

Snow Water Content	Snow pillow device and a pressure transducer	0.1 inches

Soil Moisture	Dielectric constant measuring device. Typical measurements are at 2", 4", 8", 20", and 40" where possible.	0.50%

Soil Temperature	Encapsulated thermistor. Typical measurements are at 2", 4", 8", 20", and 40" where possible.	0.1 degrees C

Solar Radiation	Pyranometer	0.01 watts per meter

Wind Speed and Direction	Propellor-type anemometer	Speed: 0.1 miles per hour; Direction: 1 degree

See the fetchSCAN tutorial for additional usage and visualization examples.

References

See the National Water and Climate Center home page for more information on the SCAN and SNOTEL programs, information on web services, and interactive maps of snow water equivalent, precipitation and streamflow.

Author

D.E. Beaudette, A.G. Brown

Examples

# \donttest{
if(requireNamespace("curl") &
    curl::has_internet()) {

    # get data
    x <- try(fetchSCAN(site.code=c(356, 2072), year=c(2015, 2016)))
    str(x)

    # get sensor metadata
    m <- SCAN_sensor_metadata(site.code=c(356, 2072))

    # get site metadata
    m <- SCAN_site_metadata(site.code=c(356, 2072))
    
    # get hourly data (396315 records)
    # x <- try(fetchSCAN(site.code=c(356, 2072), year=c(2015, 2016), timeseries = "Hourly"))
}
#> 396315 records (19.69 Mb transferred)
#> List of 14
#>  $ SMS     :'data.frame':	137883 obs. of  8 variables:
#>   ..$ Site      : int [1:137883] 356 356 356 356 356 356 356 356 356 356 ...
#>   ..$ Date      : Date[1:137883], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:137883] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:137883] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:137883] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:137883] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ depth     : num [1:137883] 5 5 5 5 5 5 5 5 5 5 ...
#>   ..$ sensor.id : Factor w/ 5 levels "SMS.I_2","SMS.I_8",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ STO     :'data.frame':	140075 obs. of  8 variables:
#>   ..$ Site      : int [1:140075] 356 356 356 356 356 356 356 356 356 356 ...
#>   ..$ Date      : Date[1:140075], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:140075] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:140075] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:140075] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:140075] 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 1.4 ...
#>   ..$ depth     : num [1:140075] 5 5 5 5 5 5 5 5 5 5 ...
#>   ..$ sensor.id : Factor w/ 5 levels "STO.I_2","STO.I_8",..: 1 1 1 1 1 1 1 1 1 1 ...
#>  $ SAL     :'data.frame':	0 obs. of  0 variables
#>  $ TAVG    :'data.frame':	17498 obs. of  8 variables:
#>   ..$ Site      : int [1:17498] 2072 2072 2072 2072 2072 2072 2072 2072 2072 2072 ...
#>   ..$ Date      : Date[1:17498], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:17498] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:17498] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:17498] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:17498] -11.9 -12.2 -12.4 -11.2 -10.8 -10.6 -10.6 -9.5 -8.9 -7.7 ...
#>   ..$ depth     : num [1:17498] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "TAVG.H": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ TMIN    :'data.frame':	17498 obs. of  8 variables:
#>   ..$ Site      : int [1:17498] 2072 2072 2072 2072 2072 2072 2072 2072 2072 2072 ...
#>   ..$ Date      : Date[1:17498], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:17498] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:17498] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:17498] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:17498] -12.2 -12.7 -12.6 -12.3 -11.1 -11 -11 -9.6 -9.1 -8.6 ...
#>   ..$ depth     : num [1:17498] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "TMIN.H": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ TMAX    :'data.frame':	17500 obs. of  8 variables:
#>   ..$ Site      : int [1:17500] 2072 2072 2072 2072 2072 2072 2072 2072 2072 2072 ...
#>   ..$ Date      : Date[1:17500], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:17500] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:17500] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:17500] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:17500] -11.7 -11.8 -12.3 -10.6 -10.4 -10.3 -9.6 -8.9 -8.6 -6.5 ...
#>   ..$ depth     : num [1:17500] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "TMAX.H": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ PRCP    :'data.frame':	13976 obs. of  8 variables:
#>   ..$ Site      : int [1:13976] 2072 2072 2072 2072 2072 2072 2072 2072 2072 2072 ...
#>   ..$ Date      : Date[1:13976], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:13976] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:13976] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:13976] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:13976] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ depth     : num [1:13976] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "PRCP.H": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ PREC    :'data.frame':	18058 obs. of  8 variables:
#>   ..$ Site      : int [1:18058] 356 356 356 356 356 356 356 356 356 356 ...
#>   ..$ Date      : Date[1:18058], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:18058] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:18058] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:18058] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:18058] 9.1 8.8 8.8 8.8 8.8 8.8 8.8 8.8 8.8 8.8 ...
#>   ..$ depth     : num [1:18058] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "PREC.I": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ SNWD    :'data.frame':	16828 obs. of  8 variables:
#>   ..$ Site      : int [1:16828] 356 356 356 356 356 356 356 356 356 356 ...
#>   ..$ Date      : Date[1:16828], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:16828] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:16828] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:16828] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:16828] 21 21 21 21 21 20 21 21 21 21 ...
#>   ..$ depth     : num [1:16828] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "SNWD.I": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ WTEQ    :'data.frame':	16999 obs. of  8 variables:
#>   ..$ Site      : int [1:16999] 356 356 356 356 356 356 356 356 356 356 ...
#>   ..$ Date      : Date[1:16999], format: "2015-01-01" "2015-01-01" ...
#>   ..$ Time      : chr [1:16999] "00:00" "01:00" "02:00" "03:00" ...
#>   ..$ water_year: num [1:16999] 2015 2015 2015 2015 2015 ...
#>   ..$ water_day : int [1:16999] 93 93 93 93 93 93 93 93 93 93 ...
#>   ..$ value     : num [1:16999] 6.1 6.2 6.2 6.2 6.2 6.2 6.2 6.2 6.3 6.2 ...
#>   ..$ depth     : num [1:16999] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ sensor.id : Factor w/ 1 level "WTEQ.I": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ WDIRV   :'data.frame':	0 obs. of  0 variables
#>  $ WSPDV   :'data.frame':	0 obs. of  0 variables
#>  $ LRADT   :'data.frame':	0 obs. of  0 variables
#>  $ metadata:'data.frame':	2 obs. of  12 variables:
#>   ..$ Name         : chr [1:2] "Blue Lakes" "Eros Data Center"
#>   ..$ Site         : num [1:2] 356 2072
#>   ..$ State        : chr [1:2] "California" "South Dakota"
#>   ..$ Network      : chr [1:2] "SNOTEL" "SCAN"
#>   ..$ County       : chr [1:2] "Alpine" "Minnehaha"
#>   ..$ Elevation_ft : num [1:2] 8067 1602
#>   ..$ Latitude     : num [1:2] 38.6 43.7
#>   ..$ Longitude    : num [1:2] -119.9 -96.6
#>   ..$ HUC          : chr [1:2] "180400120101" "101702031402"
#>   ..$ climstanm    : chr [1:2] NA "Eros Data Center"
#>   ..$ upedonid     : chr [1:2] NA "S2003SD099001"
#>   ..$ pedlabsampnum: chr [1:2] NA "03N0688"
# }