Get Ecological Dynamics Information Tool (EDIT) ecological sites by catalog (ESD/ESG) and MLRA

Data are accessed via Ecological Dynamics Interpretive Tool (EDIT) web services: https://edit.jornada.nmsu.edu/resources/esd. geoUnit refers to MLRA codes, possibly with a leading zero and trailing "X" for two digit MLRA symbols.

get_EDIT_ecoclass_by_geoUnit(geoUnit, catalog = c("esd", "esg"))

Arguments

geoUnit

A character vector of geoUnit codes e.g. c("018X","022A") for MLRAs 18 and 22A.

catalog

Catalog ID. One of: "esd" or "esg"

Value

A data.frame containing: geoUnit, id, legacyId, name. NULL if no result.

Examples

# \donttest{
 if(requireNamespace("curl") &
   curl::has_internet()) {
   get_EDIT_ecoclass_by_geoUnit(c("018X","022A"))
 }
#>    geoUnit          id    legacyId
#> 1     018X R018XI102CA R018XI102CA
#> 2     018X R018XI103CA R018XI103CA
#> 3     022A F022AC001CA F022AC001CA
#> 4     022A F022AC002CA F022AC002CA
#> 5     022A F022AC003CA F022AC003CA
#> 6     022A F022AC004CA F022AC004CA
#> 7     022A F022AC005CA F022AC005CA
#> 8     022A F022AC006CA F022AC006CA
#> 9     022A F022AC007CA F022AC007CA
#> 10    022A F022AC008CA F022AC008CA
#> 11    022A F022AE007CA F022AE007CA
#> 12    022A F022AE008CA F022AE008CA
#> 13    022A F022AE013CA F022AE013CA
#> 14    022A F022AE025CA F022AE025CA
#> 15    022A F022AF001CA F022AF001CA
#> 16    022A F022AF002CA F022AF002CA
#> 17    022A F022AF003CA F022AF003CA
#> 18    022A F022AF004CA F022AF004CA
#> 19    022A F022AF005CA F022AF005CA
#> 20    022A F022AF006CA F022AF006CA
#> 21    022A F022AX100CA F022AX100CA
#> 22    022A F022AX101CA F022AX101CA
#> 23    022A R022AA200CA R022AA200CA
#> 24    022A R022AA201CA R022AA201CA
#> 25    022A R022AC200CA R022AC200CA
#> 26    022A R022AE202CA R022AE202CA
#> 27    022A R022AE213CA R022AE213CA
#> 28    022A R022AE217CA R022AE217CA
#> 29    022A R022AX101CA R022AX101CA
#> 30    022A R022AX102CA R022AX102CA
#> 31    022A R022AX103CA R022AX103CA
#> 32    022A R022AX104CA R022AX104CA
#> 33    022A R022AX105CA R022AX105CA
#> 34    022A R022AX107CA R022AX107CA
#> 35    022A R022AZ033CA R022AZ033CA
#> 36    022A R022AZ036CA R022AZ036CA
#> 37    022A R022AZ037CA R022AZ037CA
#> 38    022A R022AZ039CA R022AZ039CA
#> 39    022A R022AZ041CA R022AZ041CA
#> 40    022A R022AZ042CA R022AZ042CA
#> 41    022A R022AZ043CA R022AZ043CA
#> 42    022A R022AZ044CA R022AZ044CA
#> 43    022A R022AZ045CA R022AZ045CA
#> 44    022A R022AZ046CA R022AZ046CA
#> 45    022A R022AZ048CA R022AZ048CA
#> 46    022A R022AZ049CA R022AZ049CA
#> 47    022A R022AZ050CA R022AZ050CA
#> 48    022A R022AZ051CA R022AZ051CA
#> 49    022A R022AZ052CA R022AZ052CA
#> 50    022A R022AZ053CA R022AZ053CA
#> 51    022A R022AZ054CA R022AZ054CA
#> 52    022A R022AZ055CA R022AZ055CA
#>                                                                                         name
#> 1  Thermic Ultramafic Foothills Extremely High Magnesium Content (Ca:Mg Ratio Less Than 0.5)
#> 2      Thermic Ultramafic Foothills Moderately High Magnesium Content (Ca:Mg Ratio 0.5 To 2)
#> 3                                                                Cryic Sandy Mountain Slopes
#> 4                                                   Cryic Sandy North Apsect Mountain Slopes
#> 5                                                                  Frigid-Cryic Sandy Slopes
#> 6                                                  Cryic Very Gravelly Loamy Mountain Slopes
#> 7                                               Cryic Sheltered, Moist Sandy Mountain Slopes
#> 8                                                           Moderately Deep Cryic Sandy Till
#> 9                                                   North-Facing Cryic Loamy Mountain Slopes
#> 10                                                            Cryic Volcanic Mountain Slopes
#> 11                                                  Frigid, Sandy, Moraines And Hill Slopes 
#> 12                                                               Frigid Loamy Moraine Slopes
#> 13                                                   Frigid, Loamy, Volcanic Mountain Slopes
#> 14                                                                       Loamy Moist Outwash
#> 15                                                  Frigid Sandy Outwash Plain Gentle Slopes
#> 16                                                           Frigid, Sandy, Or Loamy Outwash
#> 17                                                          Frigid, Loamy, Fragipan, Outwash
#> 18                                           Frigid, Shallow To Deep, Sandy Mountain Slopes 
#> 19                                    Frigid, Deep To Very Deep, Sandy-Loamy Mountain Slopes
#> 20                                                           Loamy Frigid Metamorphic Slopes
#> 21                                                         Frigid, Sandy, Moist, Outwash Fan
#> 22                                                          Moist Colluvial Headwater System
#> 23                                                                              Alpine Scree
#> 24                                                      Sandy Shallow Alpine Mountain Slopes
#> 25                                                   High Elevation Volcanic Mountain Slopes
#> 26                                                                           Granitic Pocket
#> 27                                                                        Steep Rubbly Slope
#> 28                                                                    Frigid Volcanic Slopes
#> 29                                                                 Frigid Anastomosed System
#> 30                                                                  Frigid E-C Meadow System
#> 31                                                                     Cryic E Meadow System
#> 32                                                                              Sphagnum Fen
#> 33                                                               Steep Mountain Drainageways
#> 34                                                                   Frigid C Channel System
#> 35                                                                                WET WILLOW
#> 36                                                                             MOIST CLAYPAN
#> 37                                                                                CLAY BASIN
#> 38                                                              STONY SOUTH SLOPE 16-30 P.Z.
#> 39                                                                   ERODED SLOPE 14-20 P.Z.
#> 40                                                                   SHALLOW LOAM 16-20 P.Z.
#> 41                                                                    SOUTH SLOPE 14-16 P.Z.
#> 42                                                                   COARSE LOAMY 16-20 P.Z.
#> 43                                                           GRAVELLY LOAMY SLOPE 20-30 P.Z.
#> 44                                                                             ASPEN THICKET
#> 45                                                           GRANITIC SOUTH SLOPE 14-16 P.Z.
#> 46                                                                        CLAYPAN 14-16 P.Z.
#> 47                                                                          CHANNERY MORAINE
#> 48                                                                                 KRUMMHOLZ
#> 49                                                                   GRAVELLY SLOPE 16+ P.Z.
#> 50                                                                               SNOW POCKET
#> 51                                                                      MOIST MOUNTAIN BASIN
#> 52                                                             MOUNTAIN SHOULDERS 20-30 P.Z.
# }