Get mapunit ecological sites from Soil Data Access

Get mapunit ecological sites from Soil Data Access

get_SDA_coecoclass(
  method = "None",
  areasymbols = NULL,
  mukeys = NULL,
  query_string = FALSE,
  sources = "coecosite",
  not_rated_value = "Not assigned",
  miscellaneous_areas = TRUE
)

Arguments

method
aggregation method. One of: "Dominant Component", "Dominant Condition", "None". If "None" is selected one row will be returned per component, otherwise one row will be returned per map unit.

areasymbols
vector of soil survey area symbols

mukeys
vector of map unit keys

query_string
Default: FALSE; if TRUE return a character string containing query that would be sent to SDA via SDA_query

sources
One of more of "coecosite" or "coothvegclass". If NULL no constraint on sourcesdwtablephysicalname is used in the query.

not_rated_value
Default: "Not assigned"

miscellaneous_areas
Include miscellaneous areas (non-soil components)?

Details

When method="Dominant Condition" an additional field ecoclasspct_r is returned in the result with the sum of comppct_r that have the dominant condition ecoclassid. The component with the greatest comppct_r is returned for the component and coecosite level information.

Note that if there are multiple coecoclasskey per ecoclassid there may be more than one record per component.