Get map unit properties from Soil Data Access

Get map unit properties from Soil Data Access

get_SDA_property(
  property,
  method = c("Dominant Component (Category)", "Weighted Average", "Min/Max",
    "Dominant Component (Numeric)", "Dominant Condition", "None"),
  areasymbols = NULL,
  mukeys = NULL,
  top_depth = 0,
  bottom_depth = 200,
  FUN = NULL
)

Arguments

property
a label or column name from property dictionary

method
one of: "Dominant Component (Category)", "Weighted Average", "Min/Max", "Dominant Component (Numeric)", "Dominant Condition", or "None". If "None" is selected, the number of rows returned will depend on whether a component or horizon level property was selected, otherwise the result will be 1:1 with the number of map units.

areasymbols
vector of soil survey area symbols

mukeys
vector of map unit keys

top_depth
Default: 0 (centimeters); a numeric value for upper boundary (top depth) used only for method="weighted average" and "dominant component (numeric)"

bottom_depth
Default: 200 (centimeters); a numeric value for lower boundary (bottom depth) used only for method="weighted average" and "dominant component (numeric)"

FUN
Optional: character representing SQL aggregation function either "MIN" or "MAX" for method="min/max"

Value

a data.frame with result

Details

The property argument refers to one of the property names or columns specified in the tables below.

Selected Component-level Properties

Property (Component)	Column

Range Production - Favorable Year	rsprod_h

Range Production - Normal Year	rsprod_r

Range Production - Unfavorable Year	rsprod_l

Corrosion of Steel	corsteel

Corrosion of Concrete	corcon

Drainage Class	drainagecl

Hydrologic Group	hydgrp

Taxonomic Class Name	taxclname

Taxonomic Order	taxorder

Taxonomic Suborder	taxsuborder

Taxonomic Temperature Regime	taxtempregime

Wind Erodibility Group	weg

Wind Erodibility Index	wei

t Factor	tfact

Selected Horizon-level Properties

Property (Horizon)	Column

0.1 bar H2O - Rep Value	wtenthbar_r

0.33 bar H2O - Rep Value	wthirdbar_r

15 bar H2O - Rep Value	wfifteenbar_r

Available Water Capacity - Rep Value	awc_r

Bray 1 Phosphate - Rep Value	pbray1_r

Bulk Density 0.1 bar H2O - Rep Value	dbtenthbar_r

Bulk Density 0.33 bar H2O - Rep Value	dbthirdbar_r

Bulk Density 15 bar H2O - Rep Value	dbfifteenbar_r

Bulk Density oven dry - Rep Value	dbovendry_r

CaCO3 Clay - Rep Value	claysizedcarb_r

Calcium Carbonate - Rep Value	caco3_r

Cation Exchange Capcity - Rep Value	cec7_r

Coarse Sand - Rep Value	sandco_r

Coarse Silt - Rep Value	siltco_r

Effective Cation Exchange Capcity - Rep Value	ecec_r

Electrial Conductivity 1:5 by volume - Rep Value	ec15_r

Electrical Conductivity - Rep Value	ec_r

Exchangeable Sodium Percentage - Rep Value	esp_r

Extract Aluminum - Rep Value	extral_r

Extractable Acidity - Rep Value	extracid_r

Fine Sand - Rep Value	sandfine_r

Fine Silt - Rep Value	siltfine_r

Free Iron - Rep Value	freeiron_r

Gypsum - Rep Value	gypsum_r

Kf	kffact

Ki	kifact

Kr	krfact

Kw	kwfact

LEP - Rep Value	lep_r

Liquid Limit - Rep Value	ll_r

Medium Sand - Rep Value	sandmed_r

Organic Matter - Rep Value	om_r

Oxalate Aluminum - Rep Value	aloxalate_r

Oxalate Iron - Rep Value	feoxalate_r

Oxalate Phosphate - Rep Value	poxalate_r

Plasticity Index - Rep Value	pi_r

Rock Fragments 3 - 10 cm - Rep Value	frag3to10_r

Rock Fragments > 10 cm - Rep Value	fraggt10_r

Rubbed Fiber % - Rep Value	fiberrubbedpct_r

Satiated H2O - Rep Value	wsatiated_r

Saturated Hydraulic Conductivity - Rep Value	ksat_r

Sodium Adsorption Ratio - Rep Value	sar_r

Sum of Bases - Rep Value	sumbases_r

Total Clay - Rep Value	claytotal_r

Total Phosphate - Rep Value	ptotal_r

Total Rock Fragment Volume - Rep Value	fragvoltot

Total Sand - Rep Value	sandtotal_r

Total Silt - Rep Value	silttotal_r

Unrubbed Fiber % - Rep Value	fiberunrubbedpct_r

Very Coarse Sand - Rep Value	sandvc_r

Very Fine Sand - Rep Value	sandvf_r

Water Soluble Phosphate - Rep Value	ph2osoluble_r

no. 10 sieve - Rep Value	sieveno10_r

no. 200 sieve - Rep Value	sieveno200_r

no. 4 sieve - Rep Value	sieveno4_r

no. 40 sieve - Rep Value	sieveno40_r

pH .01M CaCl2 - Rep Value	ph01mcacl2_r

pH 1:1 water - Rep Value	ph1to1h2o_r

pH Oxidized - Rep Value	phoxidized_r

Author

Jason Nemecek, Chad Ferguson, Andrew Brown