Get map unit properties from Soil Data Access

Get map unit properties from Soil Data Access

get_SDA_property(
  property,
  method = c("Dominant Component (Category)", "Weighted Average", "Min/Max",
    "Dominant Component (Numeric)", "Dominant Condition", "None"),
  areasymbols = NULL,
  mukeys = NULL,
  WHERE = NULL,
  top_depth = 0,
  bottom_depth = 200,
  FUN = NULL,
  include_minors = FALSE,
  miscellaneous_areas = FALSE,
  query_string = FALSE,
  dsn = NULL
)

Arguments

property

character vector of labels from property dictionary tables (see details) OR physical column names from component or chorizon table.

method

one of: "Dominant Component (Category)", "Dominant Component (Numeric)", "Weighted Average", "MIN", "MAX", "Dominant Condition", or "None". If "None" is selected, the number of rows returned will depend on whether a component or horizon level property was selected, otherwise the result will be 1:1 with the number of map units.

areasymbols

vector of soil survey area symbols

mukeys

vector of map unit keys

WHERE

character containing SQL WHERE clause specified in terms of fields in legend or mapunit tables, used in lieu of mukeys or areasymbols. With aggregation method "NONE" the WHERE clause may additionally contain logic for columns from the component and chorizon table.

top_depth

Default: 0 (centimeters); a numeric value for upper boundary (top depth) used only for method="Weighted Average", "Dominant Component (Numeric)", and "MIN/MAX"

bottom_depth

Default: 200 (centimeters); a numeric value for lower boundary (bottom depth) used only for method="Weighted Average", "Dominant Component (Numeric)", and "MIN/MAX"

FUN

Optional: character representing SQL aggregation function either "MIN" or "MAX" used only for method="min/max"; this argument is calculated internally if you specify method="MIN" or method="MAX"

include_minors

Include minor components in "Weighted Average" or "MIN/MAX" results?

miscellaneous_areas

Include miscellaneous areas (non-soil components) in "Weighted Average", "MIN" or "MAX" results?

query_string

Default: FALSE; if TRUE return a character string containing query that would be sent to SDA via SDA_query

dsn

Path to local SQLite database or a DBIConnection object. If NULL (default) use Soil Data Access API via SDA_query().

Value

a data.frame with result

Details

The property argument refers to one of the property names or columns specified in the tables below. Note that property can be specified as either a character vector of labeled properties, such as "Bulk Density 0.33 bar H2O - Rep Value", OR physical column names such as "dbthirdbar_r". To get "low" and "high" values for a particular property, replace the _r with _l or _h in the physical column name; for example property = c("dbthirdbar_l","dbthirdbar_r","dbthirdbar_h"). You can view exhaustive lists of component and component horizon level properties in the Soil Data Access "Tables and Columns Report".

Selected Component-level Properties

Property (Component)	Column

Range Production - Favorable Year	rsprod_h

Range Production - Normal Year	rsprod_r

Range Production - Unfavorable Year	rsprod_l

Corrosion of Steel	corsteel

Corrosion of Concrete	corcon

Drainage Class	drainagecl

Hydrologic Group	hydgrp

Taxonomic Class Name	taxclname

Taxonomic Order	taxorder

Taxonomic Suborder	taxsuborder

Taxonomic Temperature Regime	taxtempregime

Wind Erodibility Group	weg

Wind Erodibility Index	wei

t Factor	tfact

Selected Horizon-level Properties

Property (Horizon)	Column

0.1 bar H2O - Rep Value	wtenthbar_r

0.33 bar H2O - Rep Value	wthirdbar_r

15 bar H2O - Rep Value	wfifteenbar_r

Available Water Capacity - Rep Value	awc_r

Bray 1 Phosphate - Rep Value	pbray1_r

Bulk Density 0.1 bar H2O - Rep Value	dbtenthbar_r

Bulk Density 0.33 bar H2O - Rep Value	dbthirdbar_r

Bulk Density 15 bar H2O - Rep Value	dbfifteenbar_r

Bulk Density oven dry - Rep Value	dbovendry_r

CaCO3 Clay - Rep Value	claysizedcarb_r

Calcium Carbonate - Rep Value	caco3_r

Cation Exchange Capacity - Rep Value	cec7_r

Coarse Sand - Rep Value	sandco_r

Coarse Silt - Rep Value	siltco_r

Effective Cation Exchange Capacity - Rep Value	ecec_r

Electrial Conductivity 1:5 by volume - Rep Value	ec15_r

Electrical Conductivity - Rep Value	ec_r

Exchangeable Sodium Percentage - Rep Value	esp_r

Extract Aluminum - Rep Value	extral_r

Extractable Acidity - Rep Value	extracid_r

Fine Sand - Rep Value	sandfine_r

Fine Silt - Rep Value	siltfine_r

Free Iron - Rep Value	freeiron_r

Gypsum - Rep Value	gypsum_r

Kf	kffact

Ki	kifact

Kr	krfact

Kw	kwfact

LEP - Rep Value	lep_r

Liquid Limit - Rep Value	ll_r

Medium Sand - Rep Value	sandmed_r

Organic Matter - Rep Value	om_r

Oxalate Aluminum - Rep Value	aloxalate_r

Oxalate Iron - Rep Value	feoxalate_r

Oxalate Phosphate - Rep Value	poxalate_r

Plasticity Index - Rep Value	pi_r

Rock Fragments 3 - 10 inches - Rep Value	frag3to10_r

Rock Fragments > 10 inches - Rep Value	fraggt10_r

Rubbed Fiber % - Rep Value	fiberrubbedpct_r

Satiated H2O - Rep Value	wsatiated_r

Saturated Hydraulic Conductivity - Rep Value	ksat_r

Sodium Adsorption Ratio - Rep Value	sar_r

Sum of Bases - Rep Value	sumbases_r

Total Clay - Rep Value	claytotal_r

Total Phosphate - Rep Value	ptotal_r

Total Sand - Rep Value	sandtotal_r

Total Silt - Rep Value	silttotal_r

Unrubbed Fiber % - Rep Value	fiberunrubbedpct_r

Very Coarse Sand - Rep Value	sandvc_r

Very Fine Sand - Rep Value	sandvf_r

Water Soluble Phosphate - Rep Value	ph2osoluble_r

no. 10 sieve - Rep Value	sieveno10_r

no. 200 sieve - Rep Value	sieveno200_r

no. 4 sieve - Rep Value	sieveno4_r

no. 40 sieve - Rep Value	sieveno40_r

pH .01M CaCl2 - Rep Value	ph01mcacl2_r

pH 1:1 water - Rep Value	ph1to1h2o_r

pH Oxidized - Rep Value	phoxidized_r

Author

Jason Nemecek, Chad Ferguson, Andrew Brown

Examples

# \donttest{
if(requireNamespace("curl") &
    curl::has_internet()) {

 # get 1/3 bar bulk density [0,25] centimeter depth weighted average from dominant component
 get_SDA_property(property = c("dbthirdbar_l","dbthirdbar_r","dbthirdbar_h"),
                  method = "Dominant Component (Numeric)",
                  areasymbols = "CA630",
                  top_depth = 0,
                  bottom_depth = 25)
}
#> Error in curl::curl_fetch_memory(url, handle = handle): Send failure: Connection was reset
# }