Get map unit properties from Soil Data Access

Get map unit properties from Soil Data Access

get_SDA_property(
  property,
  method = c("Dominant Component (Category)", "Weighted Average", "Min/Max",
    "Dominant Component (Numeric)", "Dominant Condition", "None"),
  areasymbols = NULL,
  mukeys = NULL,
  WHERE = NULL,
  top_depth = 0,
  bottom_depth = 200,
  FUN = NULL,
  include_minors = FALSE,
  miscellaneous_areas = FALSE,
  query_string = FALSE,
  dsn = NULL
)

Arguments

property

character vector of labels from property dictionary tables (see details) OR physical column names from component or chorizon table.

method

one of: "Dominant Component (Category)", "Dominant Component (Numeric)", "Weighted Average", "MIN", "MAX", "Dominant Condition", or "None". If "None" is selected, the number of rows returned will depend on whether a component or horizon level property was selected, otherwise the result will be 1:1 with the number of map units.

areasymbols

vector of soil survey area symbols

mukeys

vector of map unit keys

WHERE

character containing SQL WHERE clause specified in terms of fields in legend or mapunit tables, used in lieu of mukeys or areasymbols. With aggregation method "NONE" the WHERE clause may additionally contain logic for columns from the component and chorizon table.

top_depth

Default: 0 (centimeters); a numeric value for upper boundary (top depth) used only for method="Weighted Average", "Dominant Component (Numeric)", and "MIN/MAX"

bottom_depth

Default: 200 (centimeters); a numeric value for lower boundary (bottom depth) used only for method="Weighted Average", "Dominant Component (Numeric)", and "MIN/MAX"

FUN

Optional: character representing SQL aggregation function either "MIN" or "MAX" used only for method="min/max"; this argument is calculated internally if you specify method="MIN" or method="MAX"

include_minors

Include minor components in "Weighted Average" or "MIN/MAX" results?

miscellaneous_areas

Include miscellaneous areas (non-soil components) in "Weighted Average", "MIN" or "MAX" results?

query_string

Default: FALSE; if TRUE return a character string containing query that would be sent to SDA via SDA_query

dsn

Path to local SQLite database or a DBIConnection object. If NULL (default) use Soil Data Access API via SDA_query().

Value

a data.frame with result

Details

The property argument refers to one of the property names or columns specified in the tables below. Note that property can be specified as either a character vector of labeled properties, such as "Bulk Density 0.33 bar H2O - Rep Value", OR physical column names such as "dbthirdbar_r". To get "low" and "high" values for a particular property, replace the _r with _l or _h in the physical column name; for example property = c("dbthirdbar_l","dbthirdbar_r","dbthirdbar_h"). You can view exhaustive lists of component and component horizon level properties in the Soil Data Access "Tables and Columns Report".

Selected Component-level Properties

Property (Component)	Column

Range Production - Favorable Year	rsprod_h

Range Production - Normal Year	rsprod_r

Range Production - Unfavorable Year	rsprod_l

Corrosion of Steel	corsteel

Corrosion of Concrete	corcon

Drainage Class	drainagecl

Hydrologic Group	hydgrp

Taxonomic Class Name	taxclname

Taxonomic Order	taxorder

Taxonomic Suborder	taxsuborder

Taxonomic Temperature Regime	taxtempregime

Wind Erodibility Group	weg

Wind Erodibility Index	wei

t Factor	tfact

Selected Horizon-level Properties

Property (Horizon)	Column

0.1 bar H2O - Rep Value	wtenthbar_r

0.33 bar H2O - Rep Value	wthirdbar_r

15 bar H2O - Rep Value	wfifteenbar_r

Available Water Capacity - Rep Value	awc_r

Bray 1 Phosphate - Rep Value	pbray1_r

Bulk Density 0.1 bar H2O - Rep Value	dbtenthbar_r

Bulk Density 0.33 bar H2O - Rep Value	dbthirdbar_r

Bulk Density 15 bar H2O - Rep Value	dbfifteenbar_r

Bulk Density oven dry - Rep Value	dbovendry_r

CaCO3 Clay - Rep Value	claysizedcarb_r

Calcium Carbonate - Rep Value	caco3_r

Cation Exchange Capacity - Rep Value	cec7_r

Coarse Sand - Rep Value	sandco_r

Coarse Silt - Rep Value	siltco_r

Effective Cation Exchange Capacity - Rep Value	ecec_r

Electrial Conductivity 1:5 by volume - Rep Value	ec15_r

Electrical Conductivity - Rep Value	ec_r

Exchangeable Sodium Percentage - Rep Value	esp_r

Extract Aluminum - Rep Value	extral_r

Extractable Acidity - Rep Value	extracid_r

Fine Sand - Rep Value	sandfine_r

Fine Silt - Rep Value	siltfine_r

Free Iron - Rep Value	freeiron_r

Gypsum - Rep Value	gypsum_r

Kf	kffact

Ki	kifact

Kr	krfact

Kw	kwfact

LEP - Rep Value	lep_r

Liquid Limit - Rep Value	ll_r

Medium Sand - Rep Value	sandmed_r

Organic Matter - Rep Value	om_r

Oxalate Aluminum - Rep Value	aloxalate_r

Oxalate Iron - Rep Value	feoxalate_r

Oxalate Phosphate - Rep Value	poxalate_r

Plasticity Index - Rep Value	pi_r

Rock Fragments 3 - 10 inches - Rep Value	frag3to10_r

Rock Fragments > 10 inches - Rep Value	fraggt10_r

Rubbed Fiber % - Rep Value	fiberrubbedpct_r

Satiated H2O - Rep Value	wsatiated_r

Saturated Hydraulic Conductivity - Rep Value	ksat_r

Sodium Adsorption Ratio - Rep Value	sar_r

Sum of Bases - Rep Value	sumbases_r

Total Clay - Rep Value	claytotal_r

Total Phosphate - Rep Value	ptotal_r

Total Sand - Rep Value	sandtotal_r

Total Silt - Rep Value	silttotal_r

Unrubbed Fiber % - Rep Value	fiberunrubbedpct_r

Very Coarse Sand - Rep Value	sandvc_r

Very Fine Sand - Rep Value	sandvf_r

Water Soluble Phosphate - Rep Value	ph2osoluble_r

no. 10 sieve - Rep Value	sieveno10_r

no. 200 sieve - Rep Value	sieveno200_r

no. 4 sieve - Rep Value	sieveno4_r

no. 40 sieve - Rep Value	sieveno40_r

pH .01M CaCl2 - Rep Value	ph01mcacl2_r

pH 1:1 water - Rep Value	ph1to1h2o_r

pH Oxidized - Rep Value	phoxidized_r

Author

Jason Nemecek, Chad Ferguson, Andrew Brown

Examples

# \donttest{
if(requireNamespace("curl") &
    curl::has_internet()) {

 # get 1/3 bar bulk density [0,25] centimeter depth weighted average from dominant component
 get_SDA_property(property = c("dbthirdbar_l","dbthirdbar_r","dbthirdbar_h"),
                  method = "Dominant Component (Numeric)",
                  areasymbols = "CA630",
                  top_depth = 0,
                  bottom_depth = 25)
}
#>     areasymbol musym
#> 1        CA630  1012
#> 2        CA630  1013
#> 3        CA630  1090
#> 4        CA630  1091
#> 5        CA630   128
#> 6        CA630   206
#> 7        CA630   207
#> 8        CA630   208
#> 9        CA630   209
#> 10       CA630   212
#> 11       CA630   220
#> 12       CA630   301
#> 13       CA630  3020
#> 14       CA630  3021
#> 15       CA630  3033
#> 16       CA630  3038
#> 17       CA630  3046
#> 18       CA630  3058
#> 19       CA630   401
#> 20       CA630  4040
#> 21       CA630  4046
#> 22       CA630  4048
#> 23       CA630  4136
#> 24       CA630  4200
#> 25       CA630  4201
#> 26       CA630  4202
#> 27       CA630   451
#> 28       CA630   475
#> 29       CA630  5012
#> 30       CA630  5013
#> 31       CA630  5015
#> 32       CA630  5016
#> 33       CA630  5051
#> 34       CA630  5053
#> 35       CA630  5057
#> 36       CA630  5100
#> 37       CA630  5101
#> 38       CA630  5201
#> 39       CA630  5202
#> 40       CA630  6029
#> 41       CA630  6033
#> 42       CA630  6034
#> 43       CA630  6036
#> 44       CA630  6037
#> 45       CA630  6038
#> 46       CA630  6039
#> 47       CA630  6041
#> 48       CA630  6043
#> 49       CA630  6054
#> 50       CA630  6055
#> 51       CA630  6070
#> 52       CA630  6071
#> 53       CA630  6072
#> 54       CA630  6074
#> 55       CA630  6075
#> 56       CA630  6076
#> 57       CA630  6078
#> 58       CA630  6079
#> 59       CA630  6202
#> 60       CA630  6205
#> 61       CA630  6206
#> 62       CA630  6207
#> 63       CA630  7065
#> 64       CA630  7066
#> 65       CA630  7074
#> 66       CA630  7076
#> 67       CA630  7078
#> 68       CA630  7079
#> 69       CA630  7083
#> 70       CA630  7085
#> 71       CA630  7086
#> 72       CA630  7087
#> 73       CA630  7088
#> 74       CA630  7089
#> 75       CA630  7091
#> 76       CA630  7092
#> 77       CA630  7096
#> 78       CA630  7155
#> 79       CA630  7156
#> 80       CA630  7159
#> 81       CA630  7165
#> 82       CA630  7166
#> 83       CA630  7207
#> 84       CA630  7210
#> 85       CA630  7211
#> 86       CA630  7212
#> 87       CA630  8026
#> 88       CA630  8033
#> 89       CA630  8034
#> 90       CA630  8036
#> 91       CA630  8110
#> 92       CA630  8111
#> 93       CA630  8115
#> 94       CA630  8120
#> 95       CA630  8160
#> 96       CA630  8161
#> 97       CA630  8162
#> 98       CA630  8171
#> 99       CA630  8172
#> 100      CA630  8173
#> 101      CA630  8175
#> 102      CA630  8176
#> 103      CA630  8177
#> 104      CA630  8178
#> 105      CA630  8190
#> 106      CA630  8194
#> 107      CA630  8286
#> 108      CA630  8287
#> 109      CA630  8289
#> 110      CA630  8302
#> 111      CA630  8307
#> 112      CA630  8312
#> 113      CA630  8314
#> 114      CA630  8317
#> 115      CA630  8318
#> 116      CA630  8319
#> 117      CA630   851
#> 118      CA630  9010
#> 119      CA630  9011
#> 120      CA630  9012
#> 121      CA630  9013
#> 122      CA630  9014
#> 123      CA630  9015
#> 124      CA630  9016
#> 125      CA630  9017
#> 126      CA630  9018
#> 127      CA630   DAM
#> 128      CA630     W
#>                                                                                               muname
#> 1                                                                                         Mined Land
#> 2                                  Mined Land-Anthraltic Xerorthents complex, 1 to 15 percent slopes
#> 3                             Ultic Haploxeralfs-Mollic Haploxeralfs complex, 3 to 30 percent slopes
#> 4                              Ultic Haploxeralfs-Aquic Dystroxerepts complex, 2 to 8 percent slopes
#> 5                                                        Cogna loam, 0 to 2 percent slopes, overwash
#> 6                                                           Pentz sandy loam, 2 to 15 percent slopes
#> 7                                                          Pentz sandy loam, 15 to 50 percent slopes
#> 8                                                     Pentz cobbly sandy loam, 2 to 8 percent slopes
#> 9                                                      Pentz-Bellota complex, 2 to 15 percent slopes
#> 10                                                                Peters clay, 2 to 8 percent slopes
#> 11                                                 Redding gravelly loam, 0 to 8 percent slopes, dry
#> 12                                          Archerdale-Hicksville association, 0 to 2 percent slopes
#> 13                                        Iron Mountain-Rock outcrop complex, 3 to 15 percent slopes
#> 14                               Iron Mountain-Crozier-Rock outcrop complex, 15 to 60 percent slopes
#> 15                              Redapple-Lilygap complex, 15 to 30 percent slopes, low precipitation
#> 16                            Devilsnose-Lilygap complex, 30 to 60 percent slopes, low precipitation
#> 17                                       Goldwall-Toomes-Rock outcrop complex, 1 to 8 percent slopes
#> 18                                            Shawsflat-Angelscreek complex, 25 to 60 percent slopes
#> 19                                                   Peters-Pentz association, 2 to 8 percent slopes
#> 20                                 Iron Mountain-Redapple-Devilsnose complex, 3 to 15 percent slopes
#> 21                                                 Redapple-Lilygap complex, 15 to 30 percent slopes
#> 22                                               Devilsnose-Lilygap complex, 30 to 60 percent slopes
#> 23  Ultic Haploxeralfs, shallow-Ultic Haploxeralfs, moderately deep complex, 10 to 35 percent slopes
#> 24                                                  Inks-Angelscreek complex, 3 to 15 percent slopes
#> 25                                                Angelscreek-Pentz complex, 15 to 30 percent slopes
#> 26                                                Angelscreek-Pentz complex, 30 to 60 percent slopes
#> 27                                                  Pentz-Peters association, 2 to 15 percent slopes
#> 28                                                  Pentz-Peters association, 2 to 50 percent slopes
#> 29                                                         Amador sandy loam, 2 to 15 percent slopes
#> 30                                               Miltonhills-Amador complex, 15 to 45 percent slopes
#> 31                                                 Ospital-Jennylind complex, 2 to 15 percent slopes
#> 32                                           Jennylind-Rock outcrop complex, 10 to 45 percent slopes
#> 33                                         Fuches-Lithic Xerorthents complex, 3 to 15 percent slopes
#> 34                                        Fuches-Lithic Xerorthents complex, 15 to 50 percent slopes
#> 35                                                                Supan loam, 5 to 30 percent slopes
#> 36        Ultic Haploxeralfs-Typic Palexerults-Aquultic haploxeralfs complex, 1 to 12 percent slopes
#> 37                   Hornitos-Red Bluff-Ultic Haploxeralfs, shallow, complex, 3 to 25 percent slopes
#> 38                                                        Pardee cobbly loam, 2 to 15 percent slopes
#> 39                                                    Pardee-Amador complex, 15 to 40 percent slopes
#> 40                                  Holland-Chawanakee-Rock outcrop complex, 45 to 90 percent slopes
#> 41                                Musick-Wukusick complex, 3 to 15 percent slopes, low precipitation
#> 42                                                   Musick-Wukusick complex, 3 to 15 percent slopes
#> 43                               Musick-Wukusick complex, 15 to 30 percent slopes, low precipitation
#> 44                                                  Musick-Wukusick complex, 15 to 30 percent slopes
#> 45                               Musick-Wukusick complex, 30 to 60 percent slopes, low precipitation
#> 46                                         Holland-Wukusick-Mantree complex, 30 to 60 percent slopes
#> 47                                                   Mantree ashy sandy loam, 3 to 15 percent slopes
#> 48                                                  Mantree ashy sandy loam, 15 to 30 percent slopes
#> 49                                         Shaver-Holland-Chawanakee complex, 8 to 30 percent slopes
#> 50                                Shaver-Lithic Humixerepts-Holland complex, 30 to 60 percent slopes
#> 51                                 Sierra-Verjeles-Aquic Haploxeralfs complex, 0 to 8 percent slopes
#> 52                                                     Sierra-Flanly complex, 3 to 15 percent slopes
#> 53                                                    Flanly-Verjeles complex, 0 to 8 percent slopes
#> 54                                                      Sierra-Orose complex, 8 to 30 percent slopes
#> 55                                                    Sierra-Flanly complex, 30 to 60 percent slopes
#> 56                                    Auberry-Hurleton-Rock outcrop complex, 20 to 60 percent slopes
#> 57                                                    Sierra-Flanly complex, 15 to 65 percent slopes
#> 58                                                               Flanly loam, 8 to 30 percent slopes
#> 59                Musick-Ultic Haploxeralfs, moderately well drained, complex, 1 to 8 percent slopes
#> 60                                                     Musick fine sandy loam, 3 to 8 percent slopes
#> 61                                                      Musick-Hotaw complex, 8 to 30 percent slopes
#> 62                                          Musick-Hotaw-Chawanakee complex, 30 to 60 percent slopes
#> 63                            Bonanza-Loafercreek complex, 3 to 15 percent slopes, low precipitation
#> 64                                   Bonanza-Loafercreek-Jasperpeak complex, 15 to 30 percent slopes
#> 65                                               Loafercreek-Bonanza complex, 3 to 15 percent slopes
#> 66                                   Bonanza-Loafercreek-Gopheridge complex, 15 to 30 percent slopes
#> 67                                            Jasperpeak-Gopheridge complex, 30 to 60 percent slopes
#> 68                                           Gopheridge-Loafercreek complex, 30 to 60 percent slopes
#> 69                           Aquic Haploxeralfs-Loafercreek-Dunstone complex, 1 to 12 percent slopes
#> 70                                               Bonanza-Loafercreek complex, 3 to 15 percent slopes
#> 71                                           Loafercreek-Gopheridge complex, 15 to 30 percent slopes
#> 72                                     Loafercreek-Gopheridge complex, cool, 15 to 30 percent slopes
#> 73                                           Loafercreek-Gopheridge complex, 30 to 60 percent slopes
#> 74                                Gardellones-Gopheridge-Motherlode complex, 30 to 60 percent slopes
#> 75                                   Trabuco-Jasperpeak-Rock outcrop complex, 8 to 30 percent slopes
#> 76                               Gopheridge-Jasperpeak-Rock outcrop complex, 30 to 60 percent slopes
#> 77                                            Gopheridge-Jasperpeak complex, 50 to 90 percent slopes
#> 78                                                Crimeahouse-Sixbit complex, 3 to 15 percent slopes
#> 79                                               Crimeahouse-Sixbit complex, 15 to 30 percent slopes
#> 80                                               Crimeahouse-Sixbit complex, 30 to 70 percent slopes
#> 81                                                Sixbit-Crimeahouse complex, 5 to 20 percent slopes
#> 82                                              Sixbit-Rock outcrop complex, 20 to 45 percent slopes
#> 83                                    Millvilla-Copperopolis-Hetchy complex, 30 to 60 percent slopes
#> 84                                                Deerflat-Millvilla complex, 3 to 15 percent slopes
#> 85                                              Millvilla-Luckymine complex, 15 to 30 percent slopes
#> 86                                             Wardsferry-Millvilla complex, 30 to 60 percent slopes
#> 87                                        Moccasinhill-Copperopolis complex, 30 to 60 percent slopes
#> 88                                      Copperopolis-Whiterock complex, 2 to 8 percent slopes, rocky
#> 89                                     Copperopolis-Whiterock complex, 3 to 15 percent slopes, rocky
#> 90                                    Copperopolis-Whiterock complex, 15 to 30 percent slopes, rocky
#> 91                                      Cumulic Humixerepts-Riverwash complex, 0 to 8 percent slopes
#> 92               Psammentic Haploxerolls-Mollic Fluvaquents-Riverwash-complex, 0 to 8 percent slopes
#> 93                                         Sanguinetti-Copperopolis complex, 30 to 60 percent slopes
#> 94                       Fluventic Haploxerepts-Oxyaquic Xerofluvents complex, 0 to 8 percent slopes
#> 95                                                   Nedsgulch-Sites complex, 3 to 15 percent slopes
#> 96                                                  Nedsgulch-Sites complex, 15 to 30 percent slopes
#> 97                                               Nedsgulch-Arpatutu complex, 30 to 60 percent slopes
#> 98                                               Nedsgulch-Wallyhill complex, 3 to 15 percent slopes
#> 99                                     Nedsgulch-Wallyhill-Arpatutu complex, 15 to 30 percent slopes
#> 100                                    Nedsgulch-Wallyhill-Arpatutu complex, 30 to 60 percent slopes
#> 101                                              Copperopolis-Hetchy complex, 8 to 30 percent slopes
#> 102                         Sanguinetti-Moccasinhill-Deerflat complex, 60 to 90 percent slopes, cool
#> 103                           Moccasinhill-Copperopolis-Sanguinetti complex, 30 to 60 percent slopes
#> 104                                       Moccasinhill-Copperopolis complex, 60 to 90 percent slopes
#> 105                                             Lickinfork-Arpatutu complex, 40 to 90 percent slopes
#> 106                                      Wallyhill, deep-Lickinfork complex, 40 to 90 percent slopes
#> 107                                                 Jocal gravelly silt loam, 8 to 30 percent slopes
#> 108                                                Jocal-Fiddletown complex, 30 to 60 percent slopes
#> 109                                         Fiddletown-Rock outcrop complex, 40 to 90 percent slopes
#> 110                                            Tiger Creek-Nedsgulch complex, 3 to 15 percent slopes
#> 111                                           Tiger Creek-Nedsgulch complex, 15 to 50 percent slopes
#> 112                                           Aquariusmine-Millvilla complex, 3 to 30 percent slopes
#> 113                     Rock outcrop-Tiger Creek-Vertic Haploxerepts complex, 1 to 45 percent slopes
#> 114                                              Beybek-Rock outcrop complex, 3 to 30 percent slopes
#> 115                                Aquariusmine-Hetchy-Rock outcrop complex, 30 to 60 percent slopes
#> 116             Mollic Haploxeralfs-Pachic Argixerolls-Rock Outcrop complex, 50 to 90 percent slopes
#> 117                                                         Mckeonhills clay, 5 to 15 percent slopes
#> 118                                                                                       Urban land
#> 119                                         Urban land-Sierra-Flanly complex, 3 to 25 percent slopes
#> 120                                          Urban land-Copperopolis complex, 0 to 15 percent slopes
#> 121                                             Urban land-Millvilla complex, 1 to 25 percent slopes
#> 122                                          Urban land-Musick-Hotaw complex, 3 to 30 percent slopes
#> 123                                  Urban land-Loafercreek-Dunstone complex, 3 to 15 percent slopes
#> 124                                  Urban land-Nedsgulch-Wallyhill complex, 15 to 30 percent slopes
#> 125                                                Urban land-Amador complex, 2 to 15 percent slopes
#> 126                                Urban land-Copperopolis-Whiterock complex, 8 to 30 percent slopes
#> 127                                                                                             Dams
#> 128                                                                                            Water
#>       mukey dbthirdbar_l dbthirdbar_r dbthirdbar_h
#> 1   2403709           NA           NA           NA
#> 2   2924914           NA           NA           NA
#> 3   2924833         1.33         1.45         1.57
#> 4   2924834         1.45         1.52         1.60
#> 5   2924907         1.43         1.48         1.54
#> 6   2924879         1.38         1.48         1.58
#> 7   2924880         1.38         1.48         1.58
#> 8   2924881         1.40         1.48         1.57
#> 9   2924882         1.35         1.43         1.50
#> 10  2924883         1.19         1.28         1.38
#> 11  2924913         1.27         1.42         1.54
#> 12  2924908         1.37         1.38         1.39
#> 13  1913607         1.30         1.41         1.52
#> 14  2425107         1.28         1.39         1.50
#> 15  2480973         0.57         0.75         0.79
#> 16  2220266         0.39         0.48         0.66
#> 17  1865918         1.36         1.42         1.49
#> 18  2450845         1.15         1.31         1.47
#> 19  2924884         1.33         1.38         1.43
#> 20  2600460         1.33         1.46         1.58
#> 21  2220270         0.57         0.75         0.79
#> 22  2220269         0.39         0.48         0.66
#> 23  2374651         1.42         1.49         1.57
#> 24  2600527         1.32         1.42         1.52
#> 25  2600528         1.11         1.36         1.50
#> 26  2600538         1.12         1.40         1.50
#> 27  2924885         1.38         1.42         1.46
#> 28  2924886         1.38         1.42         1.46
#> 29  2766838         1.50         1.56         1.62
#> 30  2600529         1.42         1.47         1.52
#> 31  2766850         1.42         1.47         1.51
#> 32  2924887         1.45         1.49         1.53
#> 33  2600465         1.36         1.47         1.57
#> 34  2600467         1.35         1.46         1.58
#> 35  2600469         0.98         1.12         1.25
#> 36  2403696         1.51         1.74         1.74
#> 37  2450478         1.40         1.47         1.53
#> 38  2766836         1.37         1.42         1.47
#> 39  2766837         1.37         1.46         1.56
#> 40  1913590         1.14         1.26         1.39
#> 41  2924832         1.10         1.17         1.24
#> 42  1913591         1.10         1.17         1.24
#> 43  2924831         1.22         1.32         1.42
#> 44  1913592         1.22         1.32         1.42
#> 45  1900697         1.37         1.45         1.53
#> 46  2482710         0.79         0.87         0.94
#> 47  2220273         0.82         0.82         1.05
#> 48  2220271         0.82         0.82         1.05
#> 49  2924756         1.21         1.28         1.36
#> 50  2924755         1.47         1.50         1.53
#> 51  2399766         1.37         1.45         1.52
#> 52  2441253         1.31         1.40         1.49
#> 53  2924738         1.40         1.50         1.60
#> 54  2399769         1.34         1.47         1.59
#> 55  2399770         1.34         1.43         1.52
#> 56  2399771         1.33         1.41         1.48
#> 57  2924701         1.38         1.46         1.54
#> 58  2924739         1.31         1.41         1.51
#> 59  2403747         1.24         1.36         1.48
#> 60  2399773         1.15         1.26         1.38
#> 61  2399774         1.33         1.43         1.53
#> 62  2399775         1.29         1.42         1.55
#> 63  3225132         1.37         1.45         1.48
#> 64  3225133         1.38         1.48         1.52
#> 65  2924751         1.28         1.40         1.52
#> 66  2924752         1.33         1.43         1.54
#> 67  2924753         1.37         1.46         1.55
#> 68  2924754         1.33         1.41         1.49
#> 69  2450844         1.24         1.37         1.51
#> 70  2441798         1.33         1.45         1.57
#> 71  2440240         1.39         1.46         1.53
#> 72  2440242         1.17         1.25         1.33
#> 73  1865926         1.31         1.38         1.46
#> 74  2450843         1.26         1.37         1.48
#> 75  3225134         1.30         1.40         1.49
#> 76  3225135         1.33         1.44         1.50
#> 77  2426483         1.38         1.44         1.50
#> 78  1865927         1.29         1.40         1.51
#> 79  1865928         1.37         1.41         1.45
#> 80  1906347         1.32         1.42         1.53
#> 81  2600534         1.36         1.45         1.54
#> 82  2766830         1.36         1.43         1.50
#> 83  1913600         1.32         1.41         1.50
#> 84  2600456         1.30         1.47         1.55
#> 85  2600457         1.32         1.44         1.56
#> 86  2600458         1.30         1.39         1.48
#> 87  2426355         1.30         1.35         1.39
#> 88  1865929         1.36         1.43         1.50
#> 89  1865930         1.36         1.43         1.50
#> 90  1865931         1.36         1.43         1.50
#> 91  2399780         1.20         1.30         1.41
#> 92  2924955         1.46         1.52         1.58
#> 93  2399783         1.24         1.36         1.46
#> 94  2924909         1.39         1.46         1.54
#> 95  1913602         1.07         1.14         1.21
#> 96  2220301         1.02         1.10         1.17
#> 97  2452459         1.05         1.16         1.27
#> 98  2383083         1.30         1.30         1.43
#> 99  2383084         1.31         1.38         1.46
#> 100 1913601         1.26         1.35         1.45
#> 101 2424962         1.31         1.38         1.45
#> 102 2424963         1.18         1.29         1.39
#> 103 1913605         1.33         1.40         1.47
#> 104 1913606         1.32         1.41         1.50
#> 105 2455490         1.29         1.37         1.46
#> 106 2483494         0.98         1.09         1.20
#> 107 2455492         1.01         1.08         1.16
#> 108 2455495         0.79         0.86         1.02
#> 109 2455494         0.82         0.92         1.02
#> 110 2436792         1.24         1.34         1.44
#> 111 2436790         1.11         1.27         1.42
#> 112 2403711         1.33         1.44         1.55
#> 113 2403710           NA           NA           NA
#> 114 2424959         1.13         1.37         1.48
#> 115 2424960         1.23         1.35         1.47
#> 116 2424961         1.34         1.44         1.54
#> 117 2924890         1.06         1.11         1.30
#> 118 2403722           NA           NA           NA
#> 119 2403721           NA           NA           NA
#> 120 2403720           NA           NA           NA
#> 121 2403719           NA           NA           NA
#> 122 2424975           NA           NA           NA
#> 123 2600480           NA           NA           NA
#> 124 2600481           NA           NA           NA
#> 125 2600537           NA           NA           NA
#> 126 2924835           NA           NA           NA
#> 127 2924912           NA           NA           NA
#> 128 2462630           NA           NA           NA
# }