Extract component data from a local NASIS Database

Extract component data from a local NASIS Database.

get_component_data_from_NASIS_db(SS = TRUE, stringsAsFactors = default.stringsAsFactors())
get_component_restrictions_from_NASIS_db(SS = TRUE)

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Dylan E. Beaudette, Stephen Roecker, and Jay M. Skovlin

See also

Examples

# \donttest{
if(local_NASIS_defined()) {
 # query text note data
 fc <- try(get_component_data_from_NASIS_db())

 # show structure of component data returned
 str(fc)
}

#> 'data.frame':	1226 obs. of  57 variables:
#>  $ dmudesc         : chr  "604101" "604101" "604101" "604101" ...
#>  $ compname        : chr  "Adinot" "Bieber" "Daphnedale" "Deven" ...
#>  $ comppct_r       : int  80 5 5 5 3 2 80 5 5 5 ...
#>  $ compkind        : chr  "series" NA NA NA ...
#>  $ majcompflag     : int  1 0 0 0 0 0 1 0 0 0 ...
#>  $ localphase      : chr  NA NA NA NA ...
#>  $ drainagecl      : chr  "moderately well" NA NA NA ...
#>  $ hydricrating    : chr  "no" "no" "no" "no" ...
#>  $ elev_l          : num  1219 NA NA NA NA ...
#>  $ elev_r          : num  1372 NA NA NA NA ...
#>  $ elev_h          : num  1524 NA NA NA NA ...
#>  $ slope_l         : num  2 NA NA NA NA NA 2 NA NA NA ...
#>  $ slope_r         : num  9 NA NA NA NA NA 9 NA NA NA ...
#>  $ slope_h         : num  15 NA NA NA NA NA 15 NA NA NA ...
#>  $ aspectccwise    : int  203 NA NA NA NA NA 158 NA NA NA ...
#>  $ aspectrep       : int  293 NA NA NA NA NA 225 NA NA NA ...
#>  $ aspectcwise     : int  23 NA NA NA NA NA 293 NA NA NA ...
#>  $ map_l           : int  305 NA NA NA NA NA 305 NA NA NA ...
#>  $ map_r           : int  356 NA NA NA NA NA 356 NA NA NA ...
#>  $ map_h           : int  406 NA NA NA NA NA 406 NA NA NA ...
#>  $ maat_l          : num  9 NA NA NA NA NA 9 NA NA NA ...
#>  $ maat_r          : num  10 NA NA NA NA NA 10 NA NA NA ...
#>  $ maat_h          : num  10 NA NA NA NA NA 10 NA NA NA ...
#>  $ mast_r          : num  NA NA NA NA NA NA NA NA NA NA ...
#>  $ reannualprecip_r: int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ ffd_l           : int  100 NA NA NA NA NA 100 NA NA NA ...
#>  $ ffd_r           : int  110 NA NA NA NA NA 110 NA NA NA ...
#>  $ ffd_h           : int  120 NA NA NA NA NA 120 NA NA NA ...
#>  $ tfact           : int  1 NA NA NA NA NA 1 NA NA NA ...
#>  $ wei             : chr  "48" NA NA NA ...
#>  $ weg             : chr  "6" NA NA NA ...
#>  $ nirrcapcl       : chr  "7" NA NA NA ...
#>  $ nirrcapscl      : chr  "s" NA NA NA ...
#>  $ nirrcapunit     : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ irrcapcl        : chr  NA NA NA NA ...
#>  $ irrcapscl       : chr  NA NA NA NA ...
#>  $ irrcapunit      : int  NA NA NA NA NA NA NA NA NA NA ...
#>  $ frostact        : chr  "moderate" NA NA NA ...
#>  $ hydricrating.1  : int  2 2 2 2 2 2 2 2 2 2 ...
#>  $ hydgrp          : chr  "d" NA NA NA ...
#>  $ corcon          : chr  "low" NA NA NA ...
#>  $ corsteel        : chr  "high" NA NA NA ...
#>  $ taxclname       : chr  "Loamy, mixed, superactive, mesic Lithic Argixerolls" NA NA NA ...
#>  $ taxorder        : chr  "mollisols" NA NA NA ...
#>  $ taxsuborder     : chr  "xerolls" NA NA NA ...
#>  $ taxgrtgroup     : chr  "argixerolls" NA NA NA ...
#>  $ taxsubgrp       : chr  "lithic argixerolls" NA NA NA ...
#>  $ taxpartsize     : chr  "loamy" NA NA NA ...
#>  $ taxpartsizemod  : chr  NA NA NA NA ...
#>  $ taxceactcl      : chr  "superactive" NA NA NA ...
#>  $ taxreaction     : chr  "not used" NA NA NA ...
#>  $ taxtempcl       : chr  "mesic" NA NA NA ...
#>  $ taxmoistscl     : chr  NA NA NA NA ...
#>  $ taxtempregime   : chr  "mesic" NA NA NA ...
#>  $ soiltaxedition  : chr  NA NA NA NA ...
#>  $ coiid           : int  694666 1687856 1687855 1687854 1687853 1687852 694667 1687887 1687888 1687889 ...
#>  $ dmuiid          : int  316001 316001 316001 316001 316001 316001 316002 316002 316002 316002 ...

# }