Extract accessory tables and summaries from a local NASIS Database

Extract accessory tables and summaries from a local NASIS Database.

get_extended_data_from_NASIS_db(SS = TRUE, nullFragsAreZero = TRUE,
                                stringsAsFactors = default.stringsAsFactors()
                                )

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

nullFragsAreZero
should fragment volumes of NULL be interpreted as 0? (default: TRUE), see details

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

See also

Examples

# \donttest{
# query extended data
e <- try(get_extended_data_from_NASIS_db())

#> Warning: some records are missing rock fragment volume, these have been removed

#> -> QC: some fragsize_h values == 76mm, may be mis-classified as cobbles [91 / 3505 records]

#> Warning: some records are missing artifact volume, these have been removed

#> Warning: all records are missing artifact volume (NULL). buffering result with NA. will be converted to zero if nullFragsAreZero = TRUE.

# show contents of extended data
str(e)

#> List of 15
#>  $ ecositehistory   :'data.frame':	0 obs. of  5 variables:
#>   ..$ siteiid        : int(0) 
#>   ..$ ecositeid      : chr(0) 
#>   ..$ ecositenm      : chr(0) 
#>   ..$ ecositecorrdate: chr(0) 
#>   ..$ es_classifier  : chr(0) 
#>  $ diagnostic       :'data.frame':	750 obs. of  4 variables:
#>   ..$ peiid   : int [1:750] 35404 35404 35404 75620 75620 75857 75857 75857 76262 109251 ...
#>   ..$ featkind: Factor w/ 84 levels "anthropic epipedon",..: 20 23 32 23 32 23 32 50 32 23 ...
#>   ..$ featdept: int [1:750] 0 23 36 0 23 0 41 119 18 0 ...
#>   ..$ featdepb: int [1:750] 36 203 178 23 183 18 165 165 183 18 ...
#>  $ diagHzBoolean    :'data.frame':	293 obs. of  21 variables:
#>   ..$ peiid                                   : int [1:293] 35404 75620 75857 76262 109251 111458 113566 115627 121340 122973 ...
#>   ..$ mollic.epipedon                         : logi [1:293] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ ochric.epipedon                         : logi [1:293] TRUE TRUE TRUE FALSE TRUE TRUE ...
#>   ..$ argillic.horizon                        : logi [1:293] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ mottles.with.chroma.2.or.less           : logi [1:293] FALSE FALSE TRUE FALSE FALSE FALSE ...
#>   ..$ kandic.horizon                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ fragipan                                : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ fragic.soil.properties                  : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ lithologic.discontinuity                : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ densic.contact                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ cambic.horizon                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ densic.materials                        : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ endosaturation                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ aquic.conditions                        : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ paralithic.contact                      : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ reduced.matrix                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ redox.concentrations                    : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ lithic.contact                          : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ redox.depletions.with.chroma.2.or.less  : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ free.carbonates                         : logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ strongly.contrasting.particle.size.class: logi [1:293] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ restriction      :'data.frame':	69 obs. of  8 variables:
#>   ..$ peiid   : int [1:69] 114228 206244 215676 215684 216189 216552 217527 217533 217649 217672 ...
#>   ..$ resdept : int [1:69] NA 40 173 103 182 150 145 163 165 167 ...
#>   ..$ resdepb : int [1:69] NA 80 203 203 203 203 203 203 203 203 ...
#>   ..$ resthk_l: int [1:69] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ resthk_r: int [1:69] NA NA 30 100 100 100 100 100 100 100 ...
#>   ..$ resthk_h: int [1:69] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ reskind : Factor w/ 21 levels "abrupt textural change",..: NA 3 1 1 1 1 1 1 1 1 ...
#>   ..$ reshard : Factor w/ 14 levels "noncemented",..: NA NA 1 1 1 1 1 1 1 1 ...
#>  $ frag_summary     :'data.frame':	3427 obs. of  18 variables:
#>   ..$ phiid               : int [1:3427] 160463 160464 160465 160466 160467 160468 160487 160488 160489 160490 ...
#>   ..$ fine_gravel         : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ cobbles             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ stones              : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:3427] 0 0 0 0 0 0 0 10 10 10 ...
#>   ..$ total_frags_pct_nopf: num [1:3427] 0 0 0 0 0 0 0 10 10 10 ...
#>   ..$ total_frags_pct     : num [1:3427] 0 0 0 0 0 0 0 10 10 10 ...
#>  $ frag_summary_v2  :'data.frame':	3440 obs. of  18 variables:
#>   ..$ phiid               : int [1:3440] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ fine_gravel         : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ cobbles             : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ stones              : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct_nopf: num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct     : num [1:3440] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ art_summary      :'data.frame':	3427 obs. of  14 variables:
#>   ..$ phiid              : int [1:3427] 160463 160464 160465 160466 160467 160468 160487 160488 160489 160490 ...
#>   ..$ art_fgr            : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_gr             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_cb             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_st             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_by             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_ch             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_fl             : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_unspecified    : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_art_pct      : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_cohesive  : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_penetrable: num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_innocuous : num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_persistent: num [1:3427] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ surf_frag_summary:'data.frame':	611 obs. of  10 variables:
#>   ..$ peiid              : int [1:611] 35313 35317 35318 35404 36266 36331 36332 37061 38225 38226 ...
#>   ..$ surface_fgravel    : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_gravel     : num [1:611] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_cobbles    : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_stones     : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_boulders   : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_channers   : num [1:611] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_flagstones : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_paragravel : logi [1:611] NA NA NA NA NA NA ...
#>   ..$ surface_paracobbles: logi [1:611] NA NA NA NA NA NA ...
#>  $ texmodifier      :'data.frame':	3624 obs. of  5 variables:
#>   ..$ peiid : int [1:3624] 35313 35313 35313 35313 35313 35313 35313 35313 35317 35317 ...
#>   ..$ phiid : int [1:3624] 160463 160464 160465 160466 160467 160468 160468 160468 160491 160491 ...
#>   ..$ phtiid: int [1:3624] 162001 162002 162003 162004 162005 162006 162007 162008 162023 162024 ...
#>   ..$ seqnum: int [1:3624] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ texmod: Factor w/ 93 levels "by","byv","byx",..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ geomorph         :'data.frame':	712 obs. of  7 variables:
#>   ..$ peiid       : int [1:712] 35404 36266 36331 36332 36332 37061 37061 38225 38225 38226 ...
#>   ..$ geomfmod    : chr [1:712] NA NA NA NA ...
#>   ..$ geomfname   : chr [1:712] "outwash plain" "outwash plain" "outwash plain" "ground moraine" ...
#>   ..$ geomfeatid  : int [1:712] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ existsonfeat: int [1:712] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ geomfiidref : int [1:712] 86 86 86 55 335 334 86 86 334 334 ...
#>   ..$ geomftname  : chr [1:712] "landform" "landform" "landform" "landform" ...
#>  $ taxhistory       :'data.frame':	1064 obs. of  20 variables:
#>   ..$ peiid         : int [1:1064] 35313 35313 35317 35317 35318 35318 35404 36266 36266 36331 ...
#>   ..$ classdate     : POSIXct[1:1064], format: "2000-04-05 00:00:00" "2012-04-06 17:11:07" ...
#>   ..$ classifier    : chr [1:1064] NA NA NA NA ...
#>   ..$ classtype     : chr [1:1064] "sampled as" "correlated" "correlated" "sampled as" ...
#>   ..$ taxonname     : chr [1:1064] "Drummer" "Drummer" "Drummer" "Drummer" ...
#>   ..$ localphase    : chr [1:1064] NA NA NA NA ...
#>   ..$ taxonkind     : chr [1:1064] NA "series" "series" NA ...
#>   ..$ seriesstatus  : chr [1:1064] NA NA NA NA ...
#>   ..$ taxpartsize   : chr [1:1064] NA "fine-silty" NA NA ...
#>   ..$ taxorder      : chr [1:1064] NA NA NA NA ...
#>   ..$ taxsuborder   : chr [1:1064] NA NA NA NA ...
#>   ..$ taxgrtgroup   : chr [1:1064] NA NA NA NA ...
#>   ..$ taxsubgrp     : chr [1:1064] NA NA NA NA ...
#>   ..$ soiltaxedition: chr [1:1064] NA NA NA NA ...
#>   ..$ osdtypelocflag: int [1:1064] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ taxmoistcl    : chr [1:1064] NA NA NA NA ...
#>   ..$ taxtempregime : chr [1:1064] NA NA NA NA ...
#>   ..$ taxfamother   : chr [1:1064] NA NA NA NA ...
#>   ..$ psctopdepth   : int [1:1064] NA NA NA NA NA NA 36 NA NA 25 ...
#>   ..$ pscbotdepth   : int [1:1064] NA NA NA NA NA NA 86 NA NA 102 ...
#>  $ photo            :'data.frame':	0 obs. of  4 variables:
#>   ..$ siteiid  : int(0) 
#>   ..$ recdate  : chr(0) 
#>   ..$ textcat  : chr(0) 
#>   ..$ imagepath: chr(0) 
#>  $ pm               :'data.frame':	544 obs. of  10 variables:
#>   ..$ siteiid     : int [1:544] 35458 35458 37117 38301 38301 38284 38592 38296 38296 38297 ...
#>   ..$ seqnum      : int [1:544] 1 2 1 1 2 1 NA 1 2 1 ...
#>   ..$ pmorder     : int [1:544] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmdept      : int [1:544] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmdepb      : int [1:544] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmmodifier  : Factor w/ 13 levels "clayey","coarse-loamy",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmgenmod    : chr [1:544] NA NA NA NA ...
#>   ..$ pmkind      : Factor w/ 187 levels "sandstone","sandstone-noncalcareous",..: 102 108 102 102 108 102 108 102 108 102 ...
#>   ..$ pmorigin    : Factor w/ 162 levels "sandstone, unspecified",..: 73 73 NA NA NA NA NA NA NA NA ...
#>   ..$ pmweathering: Factor w/ 3 levels "moderate","slight",..: NA NA NA NA NA NA NA NA NA NA ...
#>  $ struct           :'data.frame':	3359 obs. of  6 variables:
#>   ..$ phiid        : int [1:3359] 160463 160463 160464 160465 160466 160466 160467 160467 160468 160487 ...
#>   ..$ structgrade  : Factor w/ 7 levels "weak","moderate",..: 2 2 2 2 2 2 2 2 1 2 ...
#>   ..$ structsize   : Factor w/ 16 levels "coarse","coarse and very coarse",..: 6 3 3 6 6 6 6 6 6 6 ...
#>   ..$ structtype   : Factor w/ 14 levels "angular blocky",..: 1 4 4 9 8 1 8 1 8 9 ...
#>   ..$ structid     : int [1:3359] 1 2 NA NA 1 2 1 2 NA 1 ...
#>   ..$ structpartsto: int [1:3359] 2 NA NA NA 2 NA 2 NA NA 2 ...
#>  $ hzdesgn          :'data.frame':	3427 obs. of  21 variables:
#>   ..$ phiid           : int [1:3427] 160463 160464 160465 160466 160467 160468 160487 160488 160489 160490 ...
#>   ..$ seqnum          : int [1:3427] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ hzname          : chr [1:3427] "Ap" "A" "BA" "Bg1" ...
#>   ..$ hzdept          : int [1:3427] 0 20 33 43 66 89 0 25 48 71 ...
#>   ..$ hzdepb          : int [1:3427] 20 33 43 66 89 152 25 48 71 107 ...
#>   ..$ desgndisc       : int [1:3427] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ desgnmaster     : Factor w/ 67 levels "O","A","E","B",..: 2 2 4 4 4 4 2 4 4 4 ...
#>   ..$ desgnmasterprime: Factor w/ 5 levels "'","''","'''",..: NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ desgnvert       : int [1:3427] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ t               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ p               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ a               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ c               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ d               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ x               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ g               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ e               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ w               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ r               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ i               : logi [1:3427] NA NA NA NA NA NA ...
#>   ..$ b               : logi [1:3427] NA NA NA NA NA NA ...

# }