Extract accessory tables and summaries from a local NASIS Database

Extract accessory tables and summaries from a local NASIS Database.

get_extended_data_from_NASIS_db(SS = TRUE, nullFragsAreZero = TRUE,
                                stringsAsFactors = default.stringsAsFactors()
                                )

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

nullFragsAreZero
should fragment volumes of NULL be interpreted as 0? (default: TRUE), see details

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Jay M. Skovlin and Dylan E. Beaudette

See also

Examples

# \donttest{

if(local_NASIS_defined()) {
 # query extended data
 e <- try(get_extended_data_from_NASIS_db())

 # show contents of extended data
 str(e)
}

#> Warning: all records are missing rock fragment volume (NULL). buffering result with NA. will be converted to zero if nullFragsAreZero = TRUE.

#> Warning: all records are missing artifact volume (NULL). buffering result with NA. will be converted to zero if nullFragsAreZero = TRUE.

#> List of 15
#>  $ ecositehistory   :'data.frame':	0 obs. of  5 variables:
#>   ..$ siteiid        : int(0) 
#>   ..$ ecositeid      : chr(0) 
#>   ..$ ecositenm      : chr(0) 
#>   ..$ ecositecorrdate: chr(0) 
#>   ..$ es_classifier  : chr(0) 
#>  $ diagnostic       :'data.frame':	0 obs. of  4 variables:
#>   ..$ peiid   : int(0) 
#>   ..$ featkind: chr(0) 
#>   ..$ featdept: int(0) 
#>   ..$ featdepb: int(0) 
#>  $ diagHzBoolean    : NULL
#>  $ restriction      :'data.frame':	0 obs. of  8 variables:
#>   ..$ peiid   : int(0) 
#>   ..$ resdept : int(0) 
#>   ..$ resdepb : int(0) 
#>   ..$ resthk_l: int(0) 
#>   ..$ resthk_r: int(0) 
#>   ..$ resthk_h: int(0) 
#>   ..$ reskind : chr(0) 
#>   ..$ reshard : chr(0) 
#>  $ frag_summary     :'data.frame':	1 obs. of  18 variables:
#>   ..$ phiid               : int NA
#>   ..$ fine_gravel         : num 0
#>   ..$ gravel              : num 0
#>   ..$ cobbles             : num 0
#>   ..$ stones              : num 0
#>   ..$ boulders            : num 0
#>   ..$ channers            : num 0
#>   ..$ flagstones          : num 0
#>   ..$ parafine_gravel     : num 0
#>   ..$ paragravel          : num 0
#>   ..$ paracobbles         : num 0
#>   ..$ parastones          : num 0
#>   ..$ paraboulders        : num 0
#>   ..$ parachanners        : num 0
#>   ..$ paraflagstones      : num 0
#>   ..$ unspecified         : num 0
#>   ..$ total_frags_pct     : num 0
#>   ..$ total_frags_pct_nopf: num 0
#>  $ frag_summary_v2  :'data.frame':	0 obs. of  18 variables:
#>   ..$ phiid               : int(0) 
#>   ..$ fine_gravel         : logi(0) 
#>   ..$ gravel              : logi(0) 
#>   ..$ cobbles             : logi(0) 
#>   ..$ stones              : logi(0) 
#>   ..$ boulders            : logi(0) 
#>   ..$ channers            : logi(0) 
#>   ..$ flagstones          : logi(0) 
#>   ..$ parafine_gravel     : logi(0) 
#>   ..$ paragravel          : logi(0) 
#>   ..$ paracobbles         : logi(0) 
#>   ..$ parastones          : logi(0) 
#>   ..$ paraboulders        : logi(0) 
#>   ..$ parachanners        : logi(0) 
#>   ..$ paraflagstones      : logi(0) 
#>   ..$ unspecified         : logi(0) 
#>   ..$ total_frags_pct_nopf: logi(0) 
#>   ..$ total_frags_pct     : logi(0) 
#>  $ art_summary      :'data.frame':	1 obs. of  14 variables:
#>   ..$ phiid              : int NA
#>   ..$ art_fgr            : num 0
#>   ..$ art_gr             : num 0
#>   ..$ art_cb             : num 0
#>   ..$ art_st             : num 0
#>   ..$ art_by             : num 0
#>   ..$ art_ch             : num 0
#>   ..$ art_fl             : num 0
#>   ..$ art_unspecified    : num 0
#>   ..$ total_art_pct      : num 0
#>   ..$ huartvol_cohesive  : num 0
#>   ..$ huartvol_penetrable: num 0
#>   ..$ huartvol_innocuous : num 0
#>   ..$ huartvol_persistent: num 0
#>  $ surf_frag_summary:'data.frame':	0 obs. of  10 variables:
#>   ..$ peiid              : int(0) 
#>   ..$ surface_fgravel    : chr(0) 
#>   ..$ surface_gravel     : chr(0) 
#>   ..$ surface_cobbles    : chr(0) 
#>   ..$ surface_stones     : chr(0) 
#>   ..$ surface_boulders   : chr(0) 
#>   ..$ surface_channers   : chr(0) 
#>   ..$ surface_flagstones : chr(0) 
#>   ..$ surface_paragravel : chr(0) 
#>   ..$ surface_paracobbles: chr(0) 
#>  $ texmodifier      :'data.frame':	0 obs. of  5 variables:
#>   ..$ peiid : int(0) 
#>   ..$ phiid : int(0) 
#>   ..$ phtiid: int(0) 
#>   ..$ seqnum: int(0) 
#>   ..$ texmod: chr(0) 
#>  $ geomorph         :'data.frame':	0 obs. of  7 variables:
#>   ..$ peiid       : int(0) 
#>   ..$ geomfmod    : chr(0) 
#>   ..$ geomfname   : chr(0) 
#>   ..$ geomfeatid  : int(0) 
#>   ..$ existsonfeat: int(0) 
#>   ..$ geomfiidref : int(0) 
#>   ..$ geomftname  : chr(0) 
#>  $ taxhistory       :'data.frame':	0 obs. of  20 variables:
#>   ..$ peiid         : int(0) 
#>   ..$ classdate     : chr(0) 
#>   ..$ classifier    : chr(0) 
#>   ..$ classtype     : chr(0) 
#>   ..$ taxonname     : chr(0) 
#>   ..$ localphase    : chr(0) 
#>   ..$ taxonkind     : chr(0) 
#>   ..$ seriesstatus  : chr(0) 
#>   ..$ taxpartsize   : chr(0) 
#>   ..$ taxorder      : chr(0) 
#>   ..$ taxsuborder   : chr(0) 
#>   ..$ taxgrtgroup   : chr(0) 
#>   ..$ taxsubgrp     : chr(0) 
#>   ..$ soiltaxedition: chr(0) 
#>   ..$ osdtypelocflag: chr(0) 
#>   ..$ taxmoistcl    : chr(0) 
#>   ..$ taxtempregime : chr(0) 
#>   ..$ taxfamother   : chr(0) 
#>   ..$ psctopdepth   : int(0) 
#>   ..$ pscbotdepth   : int(0) 
#>  $ photo            :'data.frame':	0 obs. of  4 variables:
#>   ..$ siteiid  : int(0) 
#>   ..$ recdate  : chr(0) 
#>   ..$ textcat  : chr(0) 
#>   ..$ imagepath: chr(0) 
#>  $ pm               :'data.frame':	0 obs. of  10 variables:
#>   ..$ siteiid     : int(0) 
#>   ..$ seqnum      : int(0) 
#>   ..$ pmorder     : int(0) 
#>   ..$ pmdept      : int(0) 
#>   ..$ pmdepb      : int(0) 
#>   ..$ pmmodifier  : chr(0) 
#>   ..$ pmgenmod    : chr(0) 
#>   ..$ pmkind      : chr(0) 
#>   ..$ pmorigin    : chr(0) 
#>   ..$ pmweathering: chr(0) 
#>  $ struct           :'data.frame':	0 obs. of  6 variables:
#>   ..$ phiid        : int(0) 
#>   ..$ structgrade  : chr(0) 
#>   ..$ structsize   : chr(0) 
#>   ..$ structtype   : chr(0) 
#>   ..$ structid     : int(0) 
#>   ..$ structpartsto: int(0) 
#>  $ hzdesgn          : NULL

# }