Extract accessory tables and summaries from a local NASIS Database

Extract accessory tables and summaries from a local NASIS Database.

get_extended_data_from_NASIS_db(SS = TRUE, nullFragsAreZero = TRUE,
                                stringsAsFactors = default.stringsAsFactors()
                                )

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

nullFragsAreZero
should fragment volumes of NULL be interpreted as 0? (default: TRUE), see details

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

See also

Examples

# NOT RUN {
# query extended data
e <- get_extended_data_from_NASIS_db()

# show contents of extended data
str(e)
# }