Extract accessory tables and summaries from a local NASIS Database

Extract accessory tables and summaries from a local NASIS Database.

get_extended_data_from_NASIS_db(SS = TRUE, nullFragsAreZero = TRUE,
                                stringsAsFactors = default.stringsAsFactors()
                                )

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

nullFragsAreZero
should fragment volumes of NULL be interpreted as 0? (default: TRUE), see details

stringsAsFactors
logical: should character vectors be converted to factors? This argument is passed to the uncode() function. It does not convert those vectors that have been set outside of uncode() (i.e. hard coded). The 'factory-fresh' default is TRUE, but this can be changed by setting options(stringsAsFactors = FALSE)

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Jay M. Skovlin and Dylan E. Beaudette

See also

Examples

# \donttest{

if(local_NASIS_defined()) {
 # query extended data
 e <- try(get_extended_data_from_NASIS_db())

 # show contents of extended data
 str(e)
}

#> Warning: some records are missing rock fragment volume, these have been removed

#> Warning: fragment volume >= 100%
#> phiid:
#> 4081839
#> 677471
#> 932280
#> 932930

#> -> QC: some fragsize_h values == 76mm, may be mis-classified as cobbles [1719 / 41381 records]

#> Warning: some records are missing artifact volume, these have been removed

#> List of 15
#>  $ ecositehistory   :'data.frame':	1169 obs. of  5 variables:
#>   ..$ siteiid        : int [1:1169] 33746 48472 48480 49120 64643 108136 108264 151629 154436 154437 ...
#>   ..$ ecositeid      : chr [1:1169] "R058AC040MT" "R058CY077ND" "R058CY080ND" "R077EY058TX" ...
#>   ..$ ecositenm      : chr [1:1169] "Silty (Si)   RRU 58A-C  11-14\" p.z." "Sandy" "Loamy" "Loamy Bottomland 16-24\" PZ" ...
#>   ..$ ecositecorrdate: POSIXct[1:1169], format: NA "2018-05-08" ...
#>   ..$ es_classifier  : chr [1:1169] NA NA "Alan R. Gulsvig" NA ...
#>  $ diagnostic       :'data.frame':	12087 obs. of  4 variables:
#>   ..$ peiid   : int [1:12087] 12 12 12 12 9232 9232 9233 9233 9235 9236 ...
#>   ..$ featkind: chr [1:12087] "mollic epipedon" "argillic horizon" "calcic horizon" "lithic contact" ...
#>   ..$ featdept: int [1:12087] 0 40 90 150 0 20 0 41 0 0 ...
#>   ..$ featdepb: int [1:12087] 40 125 125 NA 20 203 41 203 33 36 ...
#>  $ diagHzBoolean    :'data.frame':	4554 obs. of  50 variables:
#>   ..$ peiid                                   : int [1:4554] 12 9232 9233 9235 9236 9274 9278 10005 10157 10205 ...
#>   ..$ mollic.epipedon                         : logi [1:4554] TRUE TRUE TRUE TRUE TRUE TRUE ...
#>   ..$ argillic.horizon                        : logi [1:4554] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ calcic.horizon                          : logi [1:4554] TRUE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ lithic.contact                          : logi [1:4554] TRUE TRUE TRUE FALSE FALSE FALSE ...
#>   ..$ cambic.horizon                          : logi [1:4554] FALSE FALSE FALSE FALSE TRUE FALSE ...
#>   ..$ free.carbonates                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ secondary.carbonates                    : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ paralithic.contact                      : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ lithologic.discontinuity                : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ gypsum.accumulations                    : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ paralithic.materials                    : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ redox.depletions.with.chroma.2.or.less  : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ umbric.epipedon                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ ochric.epipedon                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ aquic.conditions                        : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ abrupt.textural.change                  : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ duripan                                 : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ anthropic.epipedon                      : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ densic.contact                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ albic.horizon                           : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ spodic.horizon                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ andic.soil.properties                   : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ fragipan                                : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ oxic.horizon                            : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ albic.materials                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ redox.concentrations                    : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ histic.epipedon                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ sapric.soil.materials                   : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ fibric.soil.materials                   : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ hemic.soil.materials                    : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ strongly.contrasting.particle.size.class: logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ endosaturation                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ human.transported.material              : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ densic.materials                        : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ slickensides                            : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ reduced.matrix                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ ortstein                                : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ melanic.epipedon                        : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ episaturation                           : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ salt.accumulations                      : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ interfingering.of.albic.materials       : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ volcanic.glass                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ folistic.epipedon                       : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ gelic.materials                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ glossic.horizon                         : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ spodic.materials                        : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ gypsic.horizon                          : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ human.altered.material                  : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>   ..$ plaggen.epipedon                        : logi [1:4554] FALSE FALSE FALSE FALSE FALSE FALSE ...
#>  $ restriction      :'data.frame':	2584 obs. of  8 variables:
#>   ..$ peiid   : int [1:2584] 12 9232 9233 9235 9354 9354 33712 50693 51196 54900 ...
#>   ..$ resdept : int [1:2584] 150 20 41 46 20 25 66 36 27 33 ...
#>   ..$ resdepb : int [1:2584] NA 203 203 NA 25 152 91 61 51 53 ...
#>   ..$ resthk_l: int [1:2584] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ resthk_r: int [1:2584] NA 183 162 NA NA NA 25 25 NA NA ...
#>   ..$ resthk_h: int [1:2584] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ reskind : chr [1:2584] "bedrock, lithic" "bedrock, lithic" "bedrock, lithic" "bedrock, lithic" ...
#>   ..$ reshard : chr [1:2584] "indurated" "very strongly" "indurated" "indurated" ...
#>  $ frag_summary     :'data.frame':	34604 obs. of  18 variables:
#>   ..$ phiid               : int [1:34604] 23 24 28 43956 43957 43959 43960 43961 43969 43970 ...
#>   ..$ fine_gravel         : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:34604] 0 0 0 50 39 40 40 49 30 30 ...
#>   ..$ cobbles             : num [1:34604] 0 0 0 20 20 5 15 10 10 0 ...
#>   ..$ stones              : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:34604] 18 15 15 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct_nopf: num [1:34604] 18 15 15 70 59 45 55 59 40 30 ...
#>   ..$ total_frags_pct     : num [1:34604] 18 15 15 70 59 45 55 59 40 30 ...
#>  $ frag_summary_v2  :'data.frame':	34780 obs. of  18 variables:
#>   ..$ phiid               : int [1:34780] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ fine_gravel         : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ gravel              : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ cobbles             : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ stones              : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ boulders            : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ channers            : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ flagstones          : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parafine_gravel     : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paragravel          : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paracobbles         : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parastones          : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraboulders        : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ parachanners        : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ paraflagstones      : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ unspecified         : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct_nopf: num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_frags_pct     : num [1:34780] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ art_summary      :'data.frame':	34604 obs. of  14 variables:
#>   ..$ phiid              : int [1:34604] 23 24 25 26 27 28 43955 43956 43957 43958 ...
#>   ..$ art_fgr            : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_gr             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_cb             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_st             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_by             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_ch             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_fl             : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ art_unspecified    : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ total_art_pct      : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_cohesive  : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_penetrable: num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_noxious   : num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ huartvol_persistent: num [1:34604] 0 0 0 0 0 0 0 0 0 0 ...
#>  $ surf_frag_summary:'data.frame':	6668 obs. of  10 variables:
#>   ..$ peiid              : int [1:6668] 12 9232 9233 9235 9236 9274 9278 9295 9354 10005 ...
#>   ..$ surface_fgravel    : num [1:6668] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_gravel     : num [1:6668] 45 NA NA NA NA NA NA NA 50 NA ...
#>   ..$ surface_cobbles    : num [1:6668] NA NA NA NA NA NA NA NA 10 NA ...
#>   ..$ surface_stones     : num [1:6668] NA NA NA NA NA NA NA NA 10 NA ...
#>   ..$ surface_boulders   : num [1:6668] NA NA NA NA NA NA NA NA 20 NA ...
#>   ..$ surface_channers   : num [1:6668] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_flagstones : num [1:6668] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_paragravel : num [1:6668] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ surface_paracobbles: num [1:6668] NA NA NA NA NA NA NA NA NA NA ...
#>  $ texmodifier      :'data.frame':	32118 obs. of  5 variables:
#>   ..$ peiid : int [1:32118] 12 12 12 12 12 12 12 12 9232 9232 ...
#>   ..$ phiid : int [1:32118] 23 24 24 25 26 27 28 28 43956 43957 ...
#>   ..$ phtiid: int [1:32118] 21 22 22 23 24 25 26 26 44600 44601 ...
#>   ..$ seqnum: int [1:32118] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ texmod: chr [1:32118] "gr" "cnv" "ashy" NA ...
#>  $ geomorph         :'data.frame':	6891 obs. of  7 variables:
#>   ..$ peiid       : int [1:6891] 12 12 12 9232 9232 9233 9233 9235 9235 9236 ...
#>   ..$ geomfmod    : chr [1:6891] NA NA "ancient" NA ...
#>   ..$ geomfname   : chr [1:6891] "dune" "stream terrace" "river valley" "mountain slope" ...
#>   ..$ geomfeatid  : int [1:6891] 1 2 3 1 2 NA NA NA NA NA ...
#>   ..$ existsonfeat: int [1:6891] 2 3 NA 2 NA NA NA NA NA NA ...
#>   ..$ geomfiidref : int [1:6891] 40 249 328 207 324 206 324 207 324 2 ...
#>   ..$ geomftname  : chr [1:6891] "landform" "landform" "landscape" "landform" ...
#>  $ taxhistory       :'data.frame':	10605 obs. of  20 variables:
#>   ..$ peiid         : int [1:10605] 12 12 9232 9232 9233 9233 9235 9236 9236 9274 ...
#>   ..$ classdate     : POSIXct[1:10605], format: "1998-05-23 00:00:00" "2010-06-09 11:27:35" ...
#>   ..$ classifier    : chr [1:10605] NA NA NA NA ...
#>   ..$ classtype     : chr [1:10605] "sampled as" "correlated" "correlated" "sampled as" ...
#>   ..$ taxonname     : chr [1:10605] "beta" "gamma" "Brewster" "Brewster" ...
#>   ..$ localphase    : chr [1:10605] NA NA NA NA ...
#>   ..$ taxonkind     : chr [1:10605] NA "series" "series" NA ...
#>   ..$ seriesstatus  : chr [1:10605] NA NA NA NA ...
#>   ..$ taxpartsize   : chr [1:10605] NA "fine-loamy" "loamy-skeletal" NA ...
#>   ..$ taxorder      : chr [1:10605] NA "mollisols" "mollisols" NA ...
#>   ..$ taxsuborder   : chr [1:10605] NA "ustolls" "ustolls" NA ...
#>   ..$ taxgrtgroup   : chr [1:10605] NA "haplustolls" "haplustolls" NA ...
#>   ..$ taxsubgrp     : chr [1:10605] NA "aquic haplustolls" "aridic lithic haplustolls" NA ...
#>   ..$ soiltaxedition: chr [1:10605] NA NA "eighth edition" NA ...
#>   ..$ osdtypelocflag: int [1:10605] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..$ taxmoistcl    : chr [1:10605] NA NA "ustic" NA ...
#>   ..$ taxtempregime : chr [1:10605] NA "mesic" "thermic" NA ...
#>   ..$ taxfamother   : chr [1:10605] NA NA NA NA ...
#>   ..$ psctopdepth   : int [1:10605] NA 25 0 NA NA 0 0 NA 25 25 ...
#>   ..$ pscbotdepth   : int [1:10605] NA 100 20 NA NA 41 33 NA 102 102 ...
#>  $ photo            :'data.frame':	4356 obs. of  5 variables:
#>   ..$ siteiid  : int [1:4356] 307647 352131 352131 352131 352131 352131 352132 352132 352132 352132 ...
#>   ..$ recdate  : POSIXct[1:4356], format: "2008-04-21" "2009-01-16" ...
#>   ..$ textcat  : chr [1:4356] "Photo" "Photo" "Photo" "Photo" ...
#>   ..$ imagepath: chr [1:4356] "L:\\PedonPC Photos\\2008\\AMS pedons\\AMS011.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit-82.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit2_81.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit3_83.JPG" ...
#>   ..$ imagename: chr [1:4356] "AMS011.JPG" "07DWB032_pit-82.JPG" "07DWB032_pit2_81.JPG" "07DWB032_pit3_83.JPG" ...
#>  $ pm               :'data.frame':	5711 obs. of  10 variables:
#>   ..$ siteiid     : int [1:5711] 13 9228 9229 9231 9232 9271 9271 9276 9276 10023 ...
#>   ..$ seqnum      : int [1:5711] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ pmorder     : int [1:5711] NA NA NA NA NA NA NA NA NA 1 ...
#>   ..$ pmdept      : int [1:5711] NA 0 0 NA 0 NA NA NA NA NA ...
#>   ..$ pmdepb      : int [1:5711] NA 20 41 NA 203 NA NA NA NA NA ...
#>   ..$ pmmodifier  : chr [1:5711] "loamy" "gravelly" "gravelly" "gravelly" ...
#>   ..$ pmgenmod    : chr [1:5711] NA NA NA NA ...
#>   ..$ pmkind      : chr [1:5711] "alluvium" "colluvium" "colluvium" "colluvium" ...
#>   ..$ pmorigin    : chr [1:5711] NA "monzonite" "monzonite" "monzonite" ...
#>   ..$ pmweathering: chr [1:5711] NA "slight" NA "moderate" ...
#>  $ struct           :'data.frame':	25967 obs. of  6 variables:
#>   ..$ phiid        : int [1:25967] 23 24 24 25 25 26 26 28 28 43956 ...
#>   ..$ structgrade  : chr [1:25967] "strong" "weak" "moderate" "moderate" ...
#>   ..$ structsize   : chr [1:25967] "fine" "medium" "medium" "medium" ...
#>   ..$ structtype   : chr [1:25967] "granular" "granular" "subangular blocky" "granular" ...
#>   ..$ structid     : int [1:25967] NA 1 2 1 2 NA NA 1 2 NA ...
#>   ..$ structpartsto: int [1:25967] NA NA 1 NA 1 NA NA NA 1 NA ...
#>  $ hzdesgn          :'data.frame':	34604 obs. of  33 variables:
#>   ..$ phiid           : int [1:34604] 23 24 25 26 27 28 43955 43956 43957 43958 ...
#>   ..$ seqnum          : int [1:34604] 1 2 4 5 6 3 NA NA NA NA ...
#>   ..$ hzname          : chr [1:34604] "Ap" "Bt/E (Bt)" "Bt" "Btk" ...
#>   ..$ hzdept          : int [1:34604] 0 15 40 90 150 15 20 10 0 41 ...
#>   ..$ hzdepb          : int [1:34604] 15 40 90 150 NA 40 152 20 10 203 ...
#>   ..$ desgndisc       : int [1:34604] NA NA NA NA NA NA NA NA NA NA ...
#>   ..$ desgnmaster     : chr [1:34604] "A" "B/E" "B" "B" ...
#>   ..$ desgnmasterprime: chr [1:34604] NA NA NA NA ...
#>   ..$ desgnvert       : int [1:34604] NA NA NA NA NA NA NA 2 1 NA ...
#>   ..$ p               : logi [1:34604] TRUE NA FALSE FALSE NA NA ...
#>   ..$ t               : logi [1:34604] FALSE NA TRUE TRUE NA NA ...
#>   ..$ k               : logi [1:34604] FALSE NA FALSE TRUE NA NA ...
#>   ..$ w               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ r               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ b               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ y               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ z               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ i               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ ss              : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ e               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ q               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ m               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ a               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ ca              : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ s               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ x               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ c               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ h               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ d               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ g               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ j               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ o               : logi [1:34604] FALSE NA FALSE FALSE NA NA ...
#>   ..$ kk              : logi [1:34604] FALSE NA FALSE FALSE NA NA ...

# }