Extract text note data from a local NASIS Database

Extract text note data from a local NASIS Database.

get_text_notes_from_NASIS_db(SS = TRUE, fixLineEndings = TRUE)

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

fixLineEndings
convert line endings from "\r\n" to "\n"

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Dylan E. Beaudette and Jay M. Skovlin

See also

Examples

# \donttest{
if(local_NASIS_defined()) {
 # query text note data
 t <- try(get_text_notes_from_NASIS_db())

 # show contents text note data, includes: siteobs, site, pedon, horizon level text notes data.
 str(t)

 # view text categories for site text notes
 if(!inherits(t, 'try-error')) {
  table(t$site_text$textcat)
 }
}

#> List of 5
#>  $ pedon_text  :'data.frame':	37 obs. of  8 variables:
#>   ..$ recdate      : POSIXct[1:37], format: NA NA ...
#>   ..$ recauthor    : chr [1:37] NA NA NA NA ...
#>   ..$ pedontextkind: chr [1:37] "pedon conversion" "pedon note, unformatted" "pedon note, unformatted" "pedon conversion" ...
#>   ..$ textcat      : chr [1:37] "Component Name" "sitenote" "sitenote" "Component Name" ...
#>   ..$ textsubcat   : chr [1:37] NA NA NA NA ...
#>   ..$ textentry    : chr [1:37] "Peirre" "The shale is noncalcareous; when moist it is black (5Y 2/1, dry) with olive (5Y 5/3, moist) dust coat.  Parent "| __truncated__ "Vegetation:  Native grasses.  Roots mainly to 3 inches, but into shale at 38 inches. Parent Material:  Weathere"| __truncated__ "Pierre" ...
#>   ..$ peiid        : int [1:37] 105622 105622 105623 105623 105735 105736 122122 122291 122292 122293 ...
#>   ..$ petextiid    : int [1:37] 148820 144706 144707 148821 144816 144817 157207 157290 157291 157292 ...
#>  $ site_text   :'data.frame':	40 obs. of  8 variables:
#>   ..$ recdate     : POSIXct[1:40], format: "2001-09-06" "2001-09-06" ...
#>   ..$ recauthor   : chr [1:40] NA NA NA NA ...
#>   ..$ sitetextkind: chr [1:40] "conversion problem" "conversion problem" "pedon conversion" "pedon conversion" ...
#>   ..$ textcat     : chr [1:40] NA NA "Map Unit Symbol/Name" "Map Unit Symbol/Name" ...
#>   ..$ textsubcat  : chr [1:40] NA NA NA NA ...
#>   ..$ textentry   : chr [1:40] "Invalid Plant Name: sideoats grama\n\tDuplicate plant area occurrence symbols found:   BOCUC2 BOCU \n\tDuplicat"| __truncated__ "Invalid Plant Name: buffalograss\n\tDuplicate plant area occurrence symbols found:   BUCHL BUDA \n\tDuplicate p"| __truncated__ "Map Unit Symbol: PEC\nMap Unit Name:   " "Map Unit Symbol: PEC\nMap Unit Name:   " ...
#>   ..$ siteiid     : int [1:40] 105639 105752 105752 105753 105753 118509 119051 120378 120547 120548 ...
#>   ..$ sitetextiid : int [1:40] 191994 192609 199418 199419 192610 218229 218612 207989 207992 207987 ...
#>  $ siteobs_text:'data.frame':	82 obs. of  8 variables:
#>   ..$ recdate        : POSIXct[1:82], format: "2011-02-08" "2011-02-08" ...
#>   ..$ recauthor      : logi [1:82] NA NA NA NA NA NA ...
#>   ..$ siteobstextkind: chr [1:82] "miscellaneous notes" "miscellaneous notes" "site observation, unformatted" NA ...
#>   ..$ textcat        : chr [1:82] NA NA "Landform" NA ...
#>   ..$ textsubcat     : logi [1:82] NA NA NA NA NA NA ...
#>   ..$ textentry      : chr [1:82] NA NA "Site is on small slide on hillslope." NA ...
#>   ..$ site_id        : int [1:82] 579605 682064 795131 795143 509137 783826 783832 120549 782433 783958 ...
#>   ..$ siteobstextiid : int [1:82] 140215 168134 140216 NA NA NA NA NA NA NA ...
#>  $ horizon_text:'data.frame':	117 obs. of  8 variables:
#>   ..$ recdate         : POSIXct[1:117], format: "2001-09-06" "2001-09-06" ...
#>   ..$ recauthor       : chr [1:117] NA NA NA NA ...
#>   ..$ phorizontextkind: chr [1:117] "horizon note, unformatted" "horizon note, unformatted" "horizon note, unformatted" "horizon note, unformatted" ...
#>   ..$ textcat         : chr [1:117] NA NA NA NA ...
#>   ..$ textsubcat      : logi [1:117] NA NA NA NA NA NA ...
#>   ..$ textentry       : chr [1:117] "There is a distinct continuous sheen on the peds at from 5 to 18 inches. Few fine prominent red spots of weathe"| __truncated__ "Texture is shaly clay." "Texture is clay shale." "There is considerable more shale fragments than in the B2 horizon." ...
#>   ..$ phiid           : int [1:117] 514591 514592 514593 514601 514603 514996 514997 514998 514999 515000 ...
#>   ..$ phtextiid       : int [1:117] 89378 89379 89380 89381 89382 89707 89708 89709 89710 89711 ...
#>  $ photo_links :'data.frame':	0 obs. of  8 variables:
#>   ..$ recdate        : chr(0) 
#>   ..$ recauthor      : chr(0) 
#>   ..$ siteobstextkind: chr(0) 
#>   ..$ textcat        : chr(0) 
#>   ..$ textsubcat     : chr(0) 
#>   ..$ textentry      : chr(0) 
#>   ..$ site_id        : int(0) 
#>   ..$ siteobstextiid : int(0) 

#> 
#>        Form Metadata   historic grassland             Location 
#>                    2                    1                    2 
#> Map Unit Symbol/Name    native vegetation   physiographic desc 
#>                    2                    1                    1 
#>        Site location    surface fragments 
#>                    1                    1 

# }