Extract text note data from a local NASIS Database

Extract text note data from a local NASIS Database.

get_text_notes_from_NASIS_db(SS = TRUE, fixLineEndings = TRUE)

Arguments

SS
get data from the currently loaded Selected Set in NASIS or from the entire local database (default: TRUE)

fixLineEndings
convert line endings from "\r\n" to "\n"

Details

This function currently works only on Windows.

Value

A list with the results.

Author

Dylan E. Beaudette and Jay M. Skovlin

See also

Examples

# \donttest{
if(local_NASIS_defined()) {
 # query text note data
 t <- try(get_text_notes_from_NASIS_db())

 # show contents text note data, includes: siteobs, site, pedon, horizon level text notes data.
 str(t)

 # view text categories for site text notes
 if(!inherits(t, 'try-error')) {
  table(t$site_text$textcat)
 }
}

#> List of 5
#>  $ pedon_text  :'data.frame':	4961 obs. of  8 variables:
#>   ..$ recdate      : POSIXct[1:4961], format: NA NA ...
#>   ..$ recauthor    : chr [1:4961] NA NA NA NA ...
#>   ..$ pedontextkind: chr [1:4961] "pedon note, unformatted" "pedon note, unformatted" "pedon note, unformatted" "pedon note, unformatted" ...
#>   ..$ textcat      : chr [1:4961] "sitenote" "sitenote" "sitenote" "sitenote" ...
#>   ..$ textsubcat   : chr [1:4961] NA NA NA NA ...
#>   ..$ textentry    : chr [1:4961] "Correlation note and box sample number GE-88-55C" "GE-88-13G" "GE-88-25C box sample county map unit type location" "Box sample #2. Used for county type location." ...
#>   ..$ peiid        : int [1:4961] 10205 10206 10210 10237 10319 10320 10384 11090 11608 12261 ...
#>   ..$ petextiid    : int [1:4961] 16199 16200 16203 16220 16291 16292 16361 17095 17532 18117 ...
#>  $ site_text   :'data.frame':	4124 obs. of  8 variables:
#>   ..$ recdate     : POSIXct[1:4124], format: "1998-07-02 00:00:00" "2012-04-12 00:00:00" ...
#>   ..$ recauthor   : chr [1:4124] "Jim Fortner" "Jose P Quesada" "Jose P Quesada" "Jose P Quesada" ...
#>   ..$ sitetextkind: chr [1:4124] "site note, formatted" "miscellaneous notes" "miscellaneous notes" "miscellaneous notes" ...
#>   ..$ textcat     : chr [1:4124] NA NA NA NA ...
#>   ..$ textsubcat  : chr [1:4124] NA NA NA NA ...
#>   ..$ textentry   : chr [1:4124] "This site is located midway on the hillslope about 100 meters downslope of Site ID xxx." "The columns Std. Longitude, and Std. Latitude, were calculated from previous data in this Site record.\n\nThe c"| __truncated__ "The columns Std. Longitude, and Std. Latitude, were calculated from previous data in this Site record.\n\nThe c"| __truncated__ "The columns Std. Longitude, and Std. Latitude, were calculated from previous data in this Site record.\n\nThe c"| __truncated__ ...
#>   ..$ siteiid     : int [1:4124] 13 9228 9229 9232 9271 9276 9295 9360 10023 10227 ...
#>   ..$ sitetextiid : int [1:4124] 2 509399 509398 509396 509358 509354 509338 509284 644725 614325 ...
#>  $ siteobs_text:'data.frame':	10911 obs. of  8 variables:
#>   ..$ recdate        : POSIXct[1:10911], format: "2001-06-28 00:00:00" "2001-06-27 00:00:00" ...
#>   ..$ recauthor      : chr [1:10911] NA NA NA NA ...
#>   ..$ siteobstextkind: chr [1:10911] "site observation, formatted" "site observation, unformatted" "site observation, formatted" "site observation, formatted" ...
#>   ..$ textcat        : chr [1:10911] NA NA NA NA ...
#>   ..$ textsubcat     : chr [1:10911] NA NA NA NA ...
#>   ..$ textentry      : chr [1:10911] "parent material:  granitic glacial deposit." "additional vegetation:  mariposa manzanita" "Drainage:  Well drained, rapid permeability in the soil itself, but slow permeability in the \"C3r\" horizon or"| __truncated__ "additional vegetaion: mariposa manzanita.  Parent material: Granitic glacial outwash." ...
#>   ..$ site_id        : int [1:10911] 99205 99206 99207 99208 114626 114886 136454 137276 137769 139663 ...
#>   ..$ siteobstextiid : int [1:10911] 142 138 139 140 1336 338 1578 1643 1707 1821 ...
#>  $ horizon_text:'data.frame':	7498 obs. of  8 variables:
#>   ..$ recdate         : POSIXct[1:7498], format: "2005-01-11 00:00:00" "2005-01-11 00:00:00" ...
#>   ..$ recauthor       : chr [1:7498] "Wood" "Wood" "T. White" "T. White" ...
#>   ..$ phorizontextkind: chr [1:7498] "miscellaneous notes" "miscellaneous notes" "horizon note, unformatted" "horizon note, unformatted" ...
#>   ..$ textcat         : chr [1:7498] "ric" "ric" "reaction class" "reaction class" ...
#>   ..$ textsubcat      : chr [1:7498] "kind" "kind" NA NA ...
#>   ..$ textentry       : chr [1:7498] "limestone" "weathered limestone" "The Ap horizon has a neutral reaction class." "The A horizon has a neutral reaction class." ...
#>   ..$ phiid           : int [1:7498] 44473 44474 48107 48109 48108 48110 48111 48111 48112 48112 ...
#>   ..$ phtextiid       : int [1:7498] 183174 183968 968076 968077 968078 968079 968080 968081 968083 968084 ...
#>  $ photo_links :'data.frame':	4356 obs. of  8 variables:
#>   ..$ recdate        : POSIXct[1:4356], format: "2008-04-21" "2009-01-16" ...
#>   ..$ recauthor      : logi [1:4356] NA NA NA NA NA NA ...
#>   ..$ siteobstextkind: chr [1:4356] NA NA NA NA ...
#>   ..$ textcat        : chr [1:4356] "Photo" "Photo" "Photo" "Photo" ...
#>   ..$ textsubcat     : logi [1:4356] NA NA NA NA NA NA ...
#>   ..$ textentry      : chr [1:4356] "L:\\PedonPC Photos\\2008\\AMS pedons\\AMS011.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit-82.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit2_81.JPG" "L:\\PedonPCPhotos\\2007\\Dean\\Small\\07DWB032_pit3_83.JPG" ...
#>   ..$ site_id        : int [1:4356] 307647 352131 352131 352131 352131 352131 352132 352132 352132 352132 ...
#>   ..$ siteobstextiid : int [1:4356] 44535 53804 53805 53806 53807 53808 53809 53810 53811 53812 ...

#> 
#>                                         .                 102C 
#>                  389                    1                    1 
#>             31097-G8             31098-D4             31098-F1 
#>                    1                    1                    1 
#>             37120-G3                   42             42097-B5 
#>                    1                    5                    1 
#>                  43B                  80A                   85 
#>                    1                    1                    3 
#>       aerial imagery              bedrock                   CA 
#>                    1                    4                    1 
#>                CA009                CA109                CA630 
#>                    1                    3                   34 
#>              Caption   changes, additions              climate 
#>                    1                    2                    1 
#>                CO043                color          coordinates 
#>                    1                    1                    3 
#>          correlation             creation    Current MU symbol 
#>                    1                   38                    1 
#>           data entry                Depth          Depth class 
#>                    1                    1                    2 
#>          Disturbance             Drainage      ecological site 
#>                    2                    2                    3 
#>                 edit            edit note           edit notes 
#>                    1                    1                    3 
#>                edits                  ELT                  ESD 
#>                   33                    1                    2 
#>          field notes        Form Metadata              GeoDesc 
#>                    3                    3                    1 
#>          geologic MU         GPS Accuracy              History 
#>                    1                    1                    1 
#>               import          items added     items poopulated 
#>                   53                    1                    2 
#>      items populated      Items populated    lab sampling site 
#>                    2                    1                    5 
#>             Land use             Land Use             landform 
#>                    4                    4                    3 
#>            landforms              Landuse             Lat/Long 
#>                    1                    1                    1 
#>             location             Location       location edits 
#>                   11                    5                    1 
#>       Location error           manuscript Map Unit Symbol/Name 
#>                    1                    1                  260 
#>       Mapunit Number                 Misc            Misc note 
#>                    6                    1                   14 
#>        Miscellaneous         missing data                modal 
#>                    1                    3                    3 
#>                MT011          NASIS entry                NE005 
#>                    1                    1                    2 
#>        not converted               noteid                OK117 
#>                   11                   61                    1 
#>            onwership    OSD type location            ownership 
#>                    7                    1                  592 
#>      parent material      Parent material      Parent Material 
#>                    9                    1                    7 
#>     parent materials     Parent materials     Parent Materials 
#>                    1                    2                    4 
#>         permeability         Permeability   Plant Associations 
#>                    2                   17                   24 
#>         PLSS details              problem              project 
#>                    1                   12                    2 
#>  PSCS clay/rock frag             Question       representative 
#>                    6                    2                    1 
#>       rock fragments         rock outcrop         Rock outcrop 
#>                    1                    1                    4 
#>         Rock Outcrop       Rock Outcrop %              sampled 
#>                    5                    3                    6 
#>              Samples             sampling                 SDJR 
#>                    1                    1                    1 
#>         SDJR Project        sibling pedon         sister pedon 
#>                    2                    2                    2 
#>                 site                 Site  site data from memo 
#>                    4                    1                  359 
#>              site id              Site ID       Site ID Change 
#>                    3                    2                    2 
#>                slope     Slope and aspect          Slope Class 
#>                    2                    1                    1 
#>                  soi            soil temp                  STM 
#>                    5                    1                    3 
#>        Surface Cover      Surface feature                  TUD 
#>                    2                    1                    5 
#>      TUD description            TUD entry         TUD location 
#>                    1                    2                    3 
#>       TUD pedon site             TUD Site                   TX 
#>                    1                    4                    3 
#>                TX175                TX193                TX311 
#>                    2                    4                    1 
#>                TX333        type location               update 
#>                    1                    1                    2 
#>        utm, lat long           vegetation           Vegetation 
#>                    1                   12                    5 
#>                VT003                VT007                VT009 
#>                    1                   12                  517 
#>                VT011                VT025              Warning 
#>                    1                    2                    1 
#>       WGS Coordinate 
#>                    8 

# }